% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_labels}
\alias{add_labels}
\title{Axis and facet labels for plotting data.}
\usage{
add_labels(
  plot_data,
  split_factor = NULL,
  n_labs = TRUE,
  n_lab_sep = "\\n",
  labeller_fun = identity,
  n_column = "n"
)
}
\arguments{
\item{plot_data}{a data frame used for later for plotting, already pre-formatted.}

\item{split_factor}{optional, a character string with the name of the factor
defining analysis groups. If \code{split_factor = NULL}, the distributions in
the whole data frame will be shown.}

\item{n_labs}{logical. If TRUE, N numbers of complete observations are displayed
in the axis labels. Defaults to \code{TRUE}.}

\item{n_lab_sep}{separator between the category name and N number of observations.
Relevant for axis labels only if \code{n_labs = TRUE}.}

\item{labeller_fun}{a function used to transform names of variables presented
in the plot axes and facets. Defaults to identity.}

\item{n_column}{name of the data frame column storing information on N number
of complete observations to be included in axis labels if \code{n_labs = TRUE}.}
}
\value{
a data frame with additional columns \code{axis_label} and \code{facet_label}.
}
\description{
An internal function, which appends the plotting data with labels to
be presented in plot axes and facets.
}
