% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as.factor.default}
\alias{as.factor.default}
\title{Convert to a factor.}
\usage{
\method{as.factor}{default}(x)

\method{as.factor}{default}(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a factor.

a factor.
}
\description{
converts an object to a factor.

converts an object to a factor.
}
