% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as.factor.eda}
\alias{as.factor.eda}
\title{Convert numeric-type EDA to a factor.}
\usage{
\method{as.factor}{eda}(x, ...)
}
\arguments{
\item{x}{an EDA object, created by \code{\link{eda}}.}
}
\value{
a factor-type EDA.
}
\description{
converts a numeric-type EDA object to a factor-type one.
}
