% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.factor}
\alias{as.factor}
\title{Coercion to a factor.}
\usage{
as.factor(x, ...)
}
\arguments{
\item{x}{an object to be converted to a factor.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of class 'factor'.
}
\description{
Converts an object to a factor.
}
