% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_general_methods.R
\name{as.numeric.eda}
\alias{as.numeric.eda}
\alias{as.integer.eda}
\alias{factor.eda}
\alias{as.data.frame.eda}
\title{Conversion methods for \code{eda} objects.}
\usage{
\method{as.numeric}{eda}(x, ...)

\method{as.integer}{eda}(x, ...)

\method{factor}{eda}(x, levels = NULL, ...)

\method{as.data.frame}{eda}(
  x,
  row.names = NULL,
  optional = FALSE,
  col.names = "variable",
  ...
)
}
\arguments{
\item{x}{an \code{\link{eda}} object.}

\item{...}{additional arguments passed to methods}

\item{levels}{levels to be set during factor conversion.}

\item{row.names}{row names, see \code{\link[base]{as.data.frame}}.}

\item{optional}{logical, is diversion from R's naming scheme allowed?
See \code{\link[base]{as.data.frame}}}

\item{col.names}{of the data frame variable which stores
the \code{eda} object values.}
}
\description{
Conversion methods for \code{eda} objects.
}
\details{
See: \code{\link[base]{factor}}, \code{\link[base]{as.numeric}},
\code{\link[base]{as.integer}}, and \code{\link[base]{as.data.frame}}.
While converting a numeric object into a factor one, the levels corrspüond to
the sorted unique numeric values of the object.
Not that this behavior is different that the R's standard one.
}
