% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as_factor.default}
\alias{as_factor.default}
\title{Convert an object to a factor}
\usage{
\method{as_factor}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a factor.
}
\description{
Default method for converting objects to factors.
}
