% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as_numeric.default}
\alias{as_numeric.default}
\title{Convert to a numeric.}
\usage{
\method{as_numeric}{default}(x, ...)
}
\arguments{
\item{x}{an object.}
}
\value{
a numeric.
}
\description{
converts an object to a numeric.
}
