% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as_numeric.eda}
\alias{as_numeric.eda}
\title{Convert factor-type EDA to a numeric}
\usage{
\method{as_numeric}{eda}(x, ...)
}
\arguments{
\item{x}{an EDA object, created by \code{\link{eda}}.}
}
\value{
a numeric-type EDA.
}
\description{
converts a factor-type EDA object to a numeric-type one.
}
