% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_numeric}
\alias{as_numeric}
\title{Coercion to a numeric.}
\usage{
as_numeric(x, ...)
}
\arguments{
\item{x}{an object to be converted to a numeric.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of class 'numeric'.
}
\description{
Converts an object to a numeric.
}
