% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{as_table.default}
\alias{as_table.default}
\title{Default method for as_table.}
\usage{
\method{as_table}{default}(x, ...)
}
\arguments{
\item{x}{an object to be converted to a table.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of class 'table'.
}
\description{
Converts an object to a table. A wrapper around \code{\link[base]{as.table}}.
}
