% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{as_table.eda}
\alias{as_table.eda}
\title{Convert categorical EDA objects to a table.}
\usage{
\method{as_table}{eda}(eda_object, scale = c("none", "fraction", "percent"))
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.
'fraction' returns fraction of total, 'percent' returns percentages of all observations.}

\item{scale}{content of the output table: 'fraction' returns fraction of total,
'percent' returns percentages of all observations. Defaults to none.}
}
\value{
a table object.
}
\description{
converts factor-type EDA objects to a table. For numeric-type ones
NULL is returned and a warning generated.
}
