% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_table}
\alias{as_table}
\title{Tables objects.}
\usage{
as_table(x, ...)
}
\arguments{
\item{x}{an object to be converted to a table.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of class 'table'.
}
\description{
Converts an object to a table. A wrapper around \code{\link[base]{as.table}}.
}
