% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as_tibble.default}
\alias{as_tibble.default}
\title{Default method for tibble conversion}
\usage{
\method{as_tibble}{default}(x, ...)
}
\arguments{
\item{x}{an object to be converted to a tibble}
}
\value{
a tibble
}
\description{
Converts objects to a tibble.
}
