% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_tibble}
\alias{as_tibble}
\title{Convert an object to a tibble.}
\usage{
as_tibble(...)
}
\arguments{
\item{...}{extra arguments passed to methods.}

\item{x}{an object.}
}
\value{
a tibble
}
\description{
Converts an object to a tibble.
}
