% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as_vector.default}
\alias{as_vector.default}
\title{Convert to a plain vector.}
\usage{
\method{as_vector}{default}(x, ...)
}
\arguments{
\item{x}{an object to be converted to a vector.}

\item{...}{extra arguments passed to methods.}
}
\value{
a vector.
}
\description{
Converts an object to a vector.
}
