% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as_vector.eda}
\alias{as_vector.eda}
\title{Convert an EDA object to a plain vector.}
\usage{
\method{as_vector}{eda}(eda_object)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}
}
\value{
a vector with the EDA values.
}
\description{
Converts an EDA object to a vector. The data type is preserved.
}
