% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_vector}
\alias{as_vector}
\title{Coercion to a vector.}
\usage{
as_vector(x, ...)
}
\arguments{
\item{x}{an object to be converted to a vector.}

\item{...}{extra arguments passed to methods.}
}
\value{
a vector.
}
\description{
Converts an object to a vector.
}
