% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_expected_}
\alias{calc_expected_}
\title{Calculate expected values for the given variable.}
\usage{
calc_expected_(vector, fun = stats::qnorm, seed = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{vector}{a numeric vector.}

\item{fun}{a function used for generation of the distribution, \code{\link[stats]{qnorm}} by default.}

\item{seed}{seed to be set for random number generation. If NULL, no seed is set.}

\item{na.rm}{should NA's be removed from the input vector?}

\item{...}{extra arguments passed to fun.}
}
\value{
A numeric vector of the same length as vector with the expected values.
}
\description{
Calculate expected values for the given variable.
}
\details{
credits to: https://stackoverflow.com/questions/43217104/coloring-points-in-a-geom-qq-plot.
}
