% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{check_normality}
\alias{check_normality}
\alias{compare_variances}
\alias{compare_distributions}
\title{Normality and comparison of distribution of variables in a data frame.}
\usage{
check_normality(
  data,
  variables = NULL,
  split_factor = NULL,
  pub_styled = TRUE,
  ...
)

compare_variances(data, variables = NULL, split_factor, pub_styled = TRUE, ...)

compare_distributions(
  data,
  variables = NULL,
  split_factor = NULL,
  pub_styled = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector with names of variables for which the correlations
will be computed (one against all other). Note, all variables need to be numeric.
If \code{NULL}, all variables in the table except of the splitting factor are
investigated.}

\item{split_factor}{name of the splitting factor variable. For \code{check_normality()}
and \code{compare_distributions()} it can be also \code{NULL}, see Details.}

\item{pub_styled}{logical, should the output be returned as a publication-styled
data frame (class \code{\link{etest}})?}

\item{...}{additional arguments passed to the testing functions
\code{\link[fastTest]{f_shapiro_test}}, \code{\link[fastTest]{f_ks_test}} and
\code{\link[fastTest]{f_levene_test}}, and the function \code{\link{etest}}
used to format the testing results in a publication-styled manner.}
}
\value{
a data frame with the testing results: either in a native form
of the \code{fastTest} functions or, if \code{pub_styled = TRUE}, in a pre-formatted
form with significance and effect size texts, and ready-to-use plot labels.
}
\description{
Function \code{check_normality()} performs standard Shapiro-Wilk test for normality
of distribution of variables in a data frame, optionally in subset defined
by categories of a splitting factor variables.
Function \code{compare_distributions()} compares distribution of variables in a
data frame - pairwise or between two groups defined by categories of a
splitting factor variable - with two-sample asymptotic (not exact)
Kolmogorov-Smirnov test.
Function \code{compare_variances()} compares variances of variables in a data frame
between categories of a splitting factor variable with Levene test.
}
\details{
The functions employ tools of \code{fastTest} package:
\code{\link[fastTest]{f_shapiro_test}},
\code{\link[fastTest]{f_ks_test}}, and \code{\link[fastTest]{f_levene_test}}.

Please note that functions \code{check_normality()}, \code{compare_distributions()}, and
\code{compare_variances()} as exploratory tools usually used to detect violations
of assumptions of statistical tests do not offer automatic corrections of
p values for multiple testing.

\code{check_normality()}: if \code{split_factor = NULL}, normality of the distributions
is tested for the entire data frame.
\code{compare_distributions()}: if \code{split_factor = NULL}, distributions of
the variables of interest are compared in a pairwise manner.
}
