% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chi_tester}
\alias{chi_tester}
\title{Multi-sample Chi squared test.}
\usage{
chi_tester(..., test_mtx = FALSE, test_tbl = FALSE, coerce = FALSE)
}
\arguments{
\item{...}{EDA objects.}

\item{test_mtx}{logical, should a test matrix be returned instead of test results?}

\item{test_tbl}{logical, should a tibble with counts be returned instead of test results?}

\item{coerce}{logial, coerce the input to factors prior to analysis?}
}
\value{
a tibble with Chi squared, DF and p values.
}
\description{
Performs a multi-sample Chi squared test for EDA objects as specified
in (\code{\link[stats]{chisq.test}}).
}
