% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exploration.R
\name{compare}
\alias{compare}
\title{Compare or correlate features between collectives/cohorts.}
\usage{
compare(
  ...,
  variable,
  split_factor = NULL,
  what = c("test", "distribution", "variance", "correlation", "covariance", "plot"),
  type = "t_test",
  exact = TRUE,
  ci = TRUE,
  boot_method = "percentile",
  pub_styled = FALSE,
  signif_digits = 2
)
}
\arguments{
\item{...}{data frames.}

\item{variable}{a variable name.}

\item{split_factor}{optional, the name of a factor used for splitting the variables of interest into analysis groups.}

\item{what}{the requested analysis: 'test', 'distribution', 'variance', 'correlation', 'covariance'.
Defaults to 'test'.}

\item{type}{type of statistical test, see \code{\link{test}}, \code{\link{correlate}} and
\code{\link{covariance}} for details.}

\item{exact}{logical, should exact values for Chi-squared, Mann-Whitney and Wilcoxon test be returned?}

\item{ci}{logical, should confidence intervals for the test effect size be returned?}

\item{boot_method}{indicates how the bootstrap confidence intervals are calculated.
Can be any of 'percentile', 'bca', or 'normality', defaults to 'percentile'.}

\item{pub_styled}{logical, should the output be publication-ready formatted?}

\item{signif_digits}{significant digits used for rounding in the publication-style output.}
}
\description{
Compares the given variable between the provided data sets
with a range of statistical sets as described for \code{\link{test}} (statistical testing),
\code{\link{distribution}} (distribution comparison), \code{\link{variance}} (variance comparison),
\code{\link{correlate}} (correlation) or \code{\link{covariance}} (covariance).
}
\details{
in case, the split_factor is provided, only the first of the input data frames will be analyzed.
}
