% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{compare_variables}
\alias{compare_variables}
\title{Compare variable's expected values between analysis groups.}
\usage{
compare_variables(
  data,
  variables = NULL,
  split_factor,
  type = NULL,
  adj_method = "none",
  pub_styled = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a character vector of names of variables, whose expected
values will be compared between the analysis groups defined by \code{split_factor}.
If \code{variables = NULL}, all variables in the data frame except of the one
specified by \code{split_factor} will be analyzed.}

\item{split_factor}{a character string specifying the name of the splitting
factor , i.e. variable whose categories define the analysis groups.}

\item{type}{\code{NULL} or a character vector of the same length as \code{variables}
which specifies types of statistical hypothesis tests and effect size metrics.
For available tests and behavior when \code{type = NULL} see Details.}

\item{adj_method}{the method for adjusting p values for multiple testing,
as defined for \code{\link[stats]{p.adjust}},
defaults to \code{"none"}. The adjusted p value appears in the \code{p_adjusted} column.}

\item{pub_styled}{logical, should the output be returned as a publication-styled
data frame (class \code{\link{etest}})?}

\item{...}{additional arguments passed to the testing functions
and function \code{\link{etest}} used to format the results
when \code{pub_styled = TRUE}.}
}
\value{
If \code{pub_styled = FALSE}, the function returns the native output of the
testing function from \code{fastTest} package as a named list.
If \code{pub_styled = TRUE}, the function returns a data frame of class
\code{\link{etest}} storing the testing results and formatted text features
which may be used in scientific plots and tables.
}
\description{
Function \code{compare_variables}
}
\details{
The argument \code{type} specifies the type of statistical hypothesis test and
effect size statistic.
If \code{type = NULL}, distributions of factors between the analysis groups are
compared with chi-square tests with Cramer's V effect size statistics,
distributions of numeric variables are compared with Mann-Whitney tests
(two categories, biserial r effect size statistic) or Kruskal-Wallis tests
(eta-squared effect size statistic).
Available types and the corresponding testing functions are:
\itemize{
\item \code{"t"}: Student's T tests with \code{\link[fastTest]{f_t_test}}.
The effect size statistic is Cohen's d.
\item \code{"welch_t"}: Welch's T test, i.e. T test without assumption of
equality of variances, with \code{\link[fastTest]{f_t_test}}. The effect
size statistic is Cohen's d.
\item \code{"paired_t"}; paired T test with \code{\link[fastTest]{f_t_test}}.
The effect size statistic is Cohen's d.
\item \code{"wilcoxon"}: Mann-Whitney or Wilcoxon test with
\code{\link[fastTest]{f_wilcox_test}}. The effect size statistic is biserial
rank correlation coefficient r.
\item \code{"paired_wilcoxon"}: paired Wilcoxon test with \code{\link[fastTest]{f_wilcox_test}}.
The effect size statistic is biserial rank correlation coefficient r.
\item \code{"one_anova"}: one-way analysis of variance (ANOVA) with \code{\link[fastTest]{f_one_anova}}.
The effect size statistic is eta-square.
\item \code{"kruskal"}: Kruskal-Wallis test with \code{\link[fastTest]{f_kruskal_test}}.
The effect size statistic is eta-square.
\item \code{"chisq"}: chi-square test with \code{\link[fastTest]{f_chisq_test}}. The effect
size statistic is Cramer's V.
}

If \code{split_factor} variable defines more than two analysis groups and a test
comparing two categories is chosen, only the first two categories are compared
and a warning is raised.
Please note that for paired tests, the observations need to be sorted by levels
of a block-defining variable such as an identifier of an individual or a measurement
series.
The functions tests in a "safely" mode: test failures raise a warning
and are excluded from the output.
}
