% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_eda}
\alias{convert_eda}
\title{Convert multiple EDA objects to a tibble with grouping information.}
\usage{
convert_eda(..., paired = FALSE)
}
\arguments{
\item{...}{EDA objects, at least two.}

\item{paired}{logical, are the observations paired?}
}
\value{
a tibble with the factor variable 'group' and the EDAs values stored under 'variable'.
If paired is set to TRUE, the tibble contains the 'id' variable coding for sample pairing as well.
}
\description{
Tests the class and coerces multiple EDA objects to numerics.
}
