% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exploration.R
\name{correlate_variables}
\alias{correlate_variables}
\title{Correlate variables in the data set.}
\usage{
correlate_variables(
  ...,
  variables,
  what = c("correlation", "covariance"),
  type = c("pearson", "spearman", "kendall", "kappa"),
  ci = TRUE,
  boot_method = "percentile",
  pub_styled = FALSE,
  signif_digits = 2,
  adj_method = "none",
  simplify_p = TRUE
)
}
\arguments{
\item{...}{data sets.}

\item{variables}{a vector with variable names. If more than two provided, only the first two are analyzed.}

\item{what}{the requested analysis: 'correlation' or 'covariance'. Defaults to 'correlation'.}

\item{type}{the type of correlation/covariance: Pearson, Spearman, Kendall or Kohen's kappa. Defaults to Pearson.
See: \code{\link{correlate}} and \code{\link{covariance}} for details.}

\item{ci}{logical, should confidence intervals for the test effect size be returned?}

\item{boot_method}{indicates how the bootstrap confidence intervals are calculated.
Can be any of 'percentile', 'bca', or 'normality', defaults to 'percentile'.}

\item{pub_styled}{logical, should the output be publication-ready formatted?}

\item{signif_digits}{significant digits used for rounding in the publication-style output.}

\item{adj_method}{the method for adjusting p values for multiple testing, as defined for \code{\link[stats]{p.adjust}},
defaults to 'none'. The adjusted p value is summarized in the 'significance' column, when pub_style output is chosen.}

\item{simplify_p}{logical, should p_values < 0.001 be presented in a p < 0.001 form?}
}
\description{
Correlates two variables in the provided data sets by Pearson, Spearman
or Kendall method.
}
