% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corelation.R
\name{correlation}
\alias{correlation}
\title{Correlate two EDA objects.}
\usage{
correlation(
  eda_object,
  y,
  type = c("pearson", "spearman", "kendall", "kappa"),
  ci = TRUE,
  ...
)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}

\item{y}{an EDA object, created by \code{\link{eda}}.}

\item{type}{which correlation is to be calculated: Pearson (pearson), Spearman (spearman),
Kendall TauB (kendall) or Cohen's Kappa (kappa). Defaults to 'pearson'.}

\item{ci}{logical, should 95\% confidence intervals be computed?}

\item{....}{extra arguments passed to the correlation computing functions.}
}
\value{
an eTest object with test statistics, correlation coefficients and p values.
}
\description{
Correlates two EDA objects of the same length. For numeric-type EDAs,
Pearson, Spearman and Kendall TauB correlations are available. Factor-type analyses can be specifically
adressed by Cohen's kappa ('kappa'). EDA objects are converted to the respective type prior to analysis.
}
\details{
Pearson correlation is computed with \code{\link[stats]{cor.test}},
Spearman correlation is calculated with \code{\link[DescTools]{SpearmanRho}},
Kendall TauB is returned by \code{\link[DescTools]{KendallTauB}}. Kappa is calculated with
\code{\link[vcd]{Kappa}}. Correlation failures raise a warning and NULL is returned.
}
