% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{count.default}
\alias{count.default}
\title{Counting of an object occurrence.}
\usage{
\method{count}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a tibble or data frame with the category counts.
}
\description{
Counts observations in each category via \code{\link[dplyr]{count}}
}
