% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{count.eda}
\alias{count.eda}
\title{Convert categorical EDA objects to a table.}
\usage{
\method{count}{eda}(eda_object, .drop = FALSE)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}

\item{.drop}{logical, should empty levels be dropped?}
}
\value{
a tibble with the columns 'category', 'n', 'fraction' and 'percent'.
}
\description{
converts factor-type EDA objects to a table. For numeric-type ones
NULL is returned and a warning generated.
}
