% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{count}
\alias{count}
\title{Counting of an object occurrence.}
\usage{
count(x, ..., wt = NULL, sort = FALSE, name = NULL)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a tibble with the category counts.
}
\description{
Counts observations in each category.
}
