% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{cov_variables}
\alias{cov_variables}
\alias{cor_variables}
\title{Correlation and covariance coefficients.}
\usage{
cov_variables(data, variables = NULL, type = NULL, pub_styled = TRUE, ...)

cor_variables(data, variables = NULL, type = NULL, pub_styled = TRUE, ...)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector with names of variables for which the correlations
will be computed (one against all other). Note, all variables need to be of
the same type (e.g. all numeric). If \code{NULL}, all variables in the table
are correlated.}

\item{type}{\code{NULL} or a string which specifies type of co-variances or
correlations.
If \code{NULL}, Spearman's covariance or correlations are computed for numeric
variables, and Cohen's kappa for factors or integers.
Other options are \code{"pearson"}, \code{"spearman"} covariance for numeric variables,
\code{"pearson"}, \code{"spearman"}, \code{"kendallA"}, \code{"kendallB"}, \code{"xiaA"}, \code{"xiB"}
for correlation of numeric variables, and \code{"kappa"} for factors or integers.}

\item{pub_styled}{logical, should the output be returned as a publication-styled
data frame (class \code{\link{etest}})?}

\item{...}{additional arguments passed to functions computing the statistics
(\code{\link[fastTest]{f_cov}}, \code{\link[fastTest]{f_cor}},
\code{\link[fastTest]{f_kappa}}), and function \code{\link{etest}} used
to format the results when \code{pub_styled = TRUE}.}
}
\value{
If \code{pub_styled = FALSE}: the raw output of the covariance and correlation
computing functions.
If \code{pub_styled = TRUE}, a data frame of class \code{\link{etest}} with the
covariance or correlation coefficients, or Cohen's kappas is returned.
}
\description{
Pairwise correlation and covariance coefficients.
}
\details{
The function uses internally \code{\link[fastTest]{f_cov}},
\code{\link[fastTest]{f_cor}}, and \code{\link[fastTest]{f_kappa}}.
If \code{type = NULL}, \code{cov_variables()} resorts to Spearman's covariance for
numeric variables.
If \code{type = NULL}, \code{cor_variables()} resorts to Spearman's correlations
for numeric variables, and Cohen's kappa for integers and factors.
}
