% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corelation.R
\name{covariance}
\alias{covariance}
\title{Covariance of two EDA objects.}
\usage{
covariance(eda_object, y, type = c("pearson", "spearman", "kendall"), ...)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}

\item{y}{an EDA object, created by \code{\link{eda}}.}

\item{type}{which covariance is to be calculated: Pearson (pearson), Spearman (spearman),
or Kendall TauB (kendall). Defaults to 'pearson'.}

\item{....}{extra arguments passed to the correlation computing function (\code{\link[stats]{cov}}).}
}
\value{
an eTest object with the covariance coefficient.
}
\description{
Calculates covariance of two EDA objects.
For numeric-type EDAs,Pearson, Spearman and Kendall covariances are available.
Factor-type EDA objects are converted to the numeric ones prior to analysis.
}
