% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_numeric_methods.R
\name{cut.eda}
\alias{cut.eda}
\title{Cut an \code{eda} object.}
\usage{
\method{cut}{eda}(
  x,
  type = c("quartile", "mean", "median", "tertile", "custom"),
  breaks = NULL,
  labels = NULL,
  include.lowest = TRUE,
  right = TRUE,
  default_labels = TRUE,
  ...
)
}
\arguments{
\item{x}{an EDA object, created by \code{\link{eda}}.}

\item{type}{indicates how to generate cutoffs/breaks. If 'custom', the user-specified
breaks are used, otherwise the indicated statistic values are used including the
minimum and maximum values of the EDA object. Defaults to quartile.}

\item{breaks}{valid only for type 'custom'. Either a numeric vector of two or more unique
cut points or a single number (greater than or equal to 2)
giving the number of intervals into which x is to be cut.}

\item{labels}{valid only for type 'custom'. For the levels of the resulting category.
By default, labels are constructed using "(a,b]" interval notation.
If \code{labels = FALSE}, simple integer codes are returned instead of a factor.}

\item{include.lowest}{valid only for type 'custom'.
Logical, indicating if values equal to the lowest (or highest, for \code{right = FALSE})
‘breaks’ value should be included.}

\item{right}{valid only for type 'custom'. Logical, indicating if the intervals should be
closed on the right (and open on the left) or vice versa.}

\item{default_labels}{logical, should default labels be used for non-custom breaks? #
These are: Q1 to Q4 for quartiles and H1/H2 for mean and median.}

\item{...}{extra arguments passed to \code{\link[base]{cut}}.}
}
\value{
a factor \code{eda} object.
}
\description{
Converts a numeric-type EDA object to a factor-type EDA
by cutting the numeric values with the defined cutoffs.
Factor-type EDAs are returned without any changes.
}
\details{
a wrapper around \code{\link[base]{cut}}.
\code{NA} values are silently removed.
}
