% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{destat}
\alias{destat}
\alias{is_destat}
\title{Create \code{destat} objects with descriptive statistics.}
\usage{
destat(x, variables = NULL, ...)

is_destat(x)
}
\arguments{
\item{x}{a data frame with at least two column; the first one must be named
\code{variable}.}

\item{variables}{NULL or a character vector with variable names. If \code{NULL}, the variable
names will be inferred from the fist column of the \code{x} data frame.}

\item{...}{additional arguments, currently none.}
}
\value{
A data frame/tibble of \code{destat} class, which inherits most of its methods
from a "canonical" data frame.
Names of variables are stored as an attribute \code{"variable_names"}.
}
\description{
Creates a data frame of \code{destat} class, which stores variable names and
descriptive statistics in one or more analysis groups.
}
