% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corelation.R
\name{distribution}
\alias{distribution}
\title{Compare distributions of two EDA objects or with a theoretical distribution.}
\usage{
distribution(eda_object, y, ...)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}

\item{y}{form comparison of numeric distributions: an EDA object
or a numeric vector form comparing two factors: another EDA object with the same levels.}

\item{...}{extra arguments passed to \code{\link[stats]{ks.test}}.}
}
\value{
an eTest object with the D statistic and p value for numeric distribution
or an eTest object with the Chi-squared, df and p value for categorical EDAs.
}
\description{
For numeric-type EDAs: compares distributions of two EDA objects or of an EDA object with
a user-specified theoretical distribution with Kolmorgov-Smirnov test (\code{\link[stats]{ks.test}}).
For factor-type EDA objects: the distribution is compared via Chi squared test (\code{\link[stats]{chisq.test}}).
}
