% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_special_plots.R
\name{draw_freq_panel}
\alias{draw_freq_panel}
\title{Draw panel with frequencies or counts}
\usage{
draw_freq_panel(
  data,
  variables,
  split_factor = NULL,
  rm_na = TRUE,
  scale = c("percent", "fraction", "count"),
  form = c("stack", "dodge"),
  show_labels = FALSE,
  txt_size = 2.75,
  geom_label = TRUE,
  signif_digits = 2,
  alpha = 1,
  dodge_w = 0.9,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "\% of cohort",
  y_lab = split_factor,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{names of the variables to plot, they need to be factors.}

\item{split_factor}{name of the splitting variable.}

\item{rm_na}{logical, should missing observations be removed prior
to plotting? If FALSE,}

\item{scale}{statistic to be plotted: percent (default), fraction or counts
of observations for each variable level.}

\item{form}{form of the bar plot: stacked bars (default) or dodged bars.}

\item{show_labels}{logical, should the data bars be labeled with the percent,
fraction or count value?}

\item{txt_size}{size of the label text, ignored if show_labels is FALSE.}

\item{signif_digits}{significant digits to be shown in the bar labels.}

\item{alpha}{alpha of the bar or distribution width ribbon.}

\item{dodge_w}{dodge width.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{x_lab}{x axis title.}

\item{y_lab}{Y axis title.}

\item{cust_theme}{custom ggplot theme.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{facet_grid}}.
Ignored, if split_factor is NULL.}

\item{split}{name of the splitting variable (optional). If provided,
the panel will be faceted by the variables and the split_factor presented
in the Y axis.}

\item{geom_labe}{logical, should the bar label be presented as a geom_label?
If FALSE, plain text is displayed.}
}
\value{
a ggplot object.
}
\description{
Draws a panel of stacked or dodged bar plots with frequencies
or counts for the given categorical variables. Works best, when the variables
share the same levels.
}
