% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_plots.R
\name{draw_numeric_panel}
\alias{draw_numeric_panel}
\title{Panel with distribution of multiple variables in a data frame.}
\usage{
draw_numeric_panel(
  data,
  variables,
  split_factor = NULL,
  distr_geom = c("violin", "box"),
  alpha = 0.25,
  non_zero = FALSE,
  point_size = 2,
  point_rim_color = "black",
  point_alpha = 0.5,
  point_hjitter = 0.15,
  point_wjitter = 0,
  dodge_w = 0.75,
  fill = "cornsilk4",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "Variable value",
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a character vector with variable names}

\item{split_factor}{name of the data frame's variable, whose categories will
be used to split values of \code{variable} in the plot. Provided as a string.}

\item{point_size}{size of data points}

\item{point_alpha}{alpha (opacity) of data points.}

\item{point_hjitter}{horizontal jittering of data points}

\item{point_wjitter}{vertical jittering of data points.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle; if \code{NULL} numbers of all and complete
observations are shown here.}

\item{x_lab}{title of the X axis.}

\item{cust_theme}{custom \code{ggplot}'s \code{theme} object or \code{NULL}. If \code{NULL} no
custom theme is applied, which makes the plot objects much smaller.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_boxplot}},
\code{\link[ggplot2]{geom_violin}}, \code{\link[ggplot2]{geom_histogram}}.}
}
\value{
a \code{ggplot} object.
}
\description{
\code{ggplot} graphics with plots of distribution of multiple numeric or factor
variables in a data frame.
}
