% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_special_plots.R
\name{draw_quantile_elli}
\alias{draw_quantile_elli}
\title{Draw data quantile ellipses.}
\usage{
draw_quantile_elli(
  data,
  variables,
  split_factor = NULL,
  quantiles = c(0.25, 0.75),
  non_zero = FALSE,
  ell_width = 0.5,
  alpha = 0.5,
  fill = "cornsilk4",
  median_color = "coral3",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "median with IQR",
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{names of the variables to plot.}

\item{quantiles}{quantiles to be shown on the longer axis (a dimension)
of ellipses, a two element vector.}

\item{non_zero}{logical, should the zero values be removed?}

\item{ell_width}{length of the shorter ellipse axis (b dimension).}

\item{alpha}{alpha of the ellipses.}

\item{fill}{fill color of the ellipses, ignored when split_factor is provided.}

\item{median_color}{color of the median points,
ignored when split factor provided.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{x_lab}{x axis title.}

\item{cust_theme}{custom ggplot theme.}

\item{...}{extra arguments passed to \code{\link[ggforce]{stat_ellip}}.}

\item{spit_factor}{name of the splitting variable.}
}
\description{
Draws median (diamond) and user-specified data quantiles
(ellipse).
}
\details{
employs \code{\link[ggforce]{stat_ellip}}.
}
