% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_special_plots.R
\name{draw_stat_panel}
\alias{draw_stat_panel}
\title{Draw a panel of statistic values.}
\usage{
draw_stat_panel(
  data,
  variables,
  split_factor = NULL,
  stat = c("mean", "median"),
  err_stat = c("none", "sd", "se", "2se", "iqr", "95perc"),
  form = c("bar", "line"),
  alpha = 0.5,
  fill = "cornsilk4",
  dodge_w = if (form == "bar") 0.9 else 0,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "Mean",
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{names of the variables to plot.}

\item{split_factor}{name of the splitting variable.}

\item{stat}{distribution statistic, currently mean or median supported.}

\item{err_stat}{error/distribution width statistic, SD ('sd'), SEM ('se'),
95% CI from the normal distribution (2se), IQR (iqr) or 95% percentile
('95perc') are currently supported.}

\item{form}{for of the plot: bar (default) or line.}

\item{alpha}{alpha of the bar or distribution width ribbon.}

\item{fill}{fill color of the violins and data points,
ignored when split_factor is provided.}

\item{dodge_w}{dodge width.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{x_lab}{x axis title.}

\item{cust_theme}{custom ggplot theme.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{geom_ribbon}}}
}
\value{
a ggplot object.
}
\description{
Draws a panel of specified distribution statistic (e.g. median
or mean) in form of bars or lines.
}
