% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_special_plots.R
\name{draw_violin_panel}
\alias{draw_violin_panel}
\title{Draw a panel of violin or box plots.}
\usage{
draw_violin_panel(
  data,
  variables,
  split_factor = NULL,
  distr_geom = c("violin", "box"),
  non_zero = FALSE,
  point_alpha = 0.5,
  point_hjitter = 0.15,
  point_wjitter = 0,
  dodge_w = 0.75,
  fill = "cornsilk4",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "Variable value",
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{names of the variables to plot.}

\item{distr_geom}{form of the distribution symbol: a violin or box plot.}

\item{non_zero}{logical, should the zero values be removed?}

\item{point_alpha}{alpha of data points.}

\item{point_hjitter}{point jitter height.}

\item{point_wjitter}{point jitter width.}

\item{dodge_w}{dodge width.}

\item{fill}{fill color of the violins and data points,
ignored when split_factor is provided.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{x_lab}{x axis title.}

\item{cust_theme}{custom ggplot theme.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{geom_violin}} or
\code{\link[ggplot2]{geom_box}}.}

\item{spit_factor}{name of the splitting variable.}
}
\description{
Draws violin  or box plots for the specified variable set.
Medians and IQR are presented as orange diamonds with whiskers.
Points represent single observations.
}
