% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{eda.character}
\alias{eda.character}
\title{Generate EDA objects from character vectors.}
\usage{
\method{eda}{character}(x)
}
\arguments{
\item{x}{an object to be converted to eda.}
}
\value{
an EDA object.
}
\description{
Generates EDA objects from a character vector.
}
