% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{eda.data.frame}
\alias{eda.data.frame}
\title{Generate an EDA object from a data frame variable.}
\usage{
\method{eda}{data.frame}(data, variable, ...)
}
\arguments{
\item{data}{and object to be converted to EDA.}

\item{variable}{a name or reference (quoted or unquoted) to a data frame variable.}
}
\value{
an EDA object.
}
\description{
Generates EDA objects from a data frame variable.
}
