% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{eda.default}
\alias{eda.default}
\title{Generate EDA objects.}
\usage{
\method{eda}{default}(x, ...)
}
\arguments{
\item{x}{and object to be converted to EDA.}
}
\value{
an EDA object.
}
\description{
Generates EDA objects from multiple inputs such as vectors or data frame variables.
}
