% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{eda.factor}
\alias{eda.factor}
\title{Generate EDA objects from factors.}
\usage{
\method{eda}{factor}(x)
}
\arguments{
\item{x}{an object to be converted to EDA.}
}
\value{
an EDA object.
}
\description{
Generates EDA objects from a factor.
}
