% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{eda.logical}
\alias{eda.logical}
\title{Generate EDA objects from logical vectors.}
\usage{
\method{eda}{logical}(x, ...)
}
\arguments{
\item{x}{an object to be converted to EDA.}
}
\value{
an EDA object.
}
\description{
Generates EDA objects from a logical vector.
}
