% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{eda.numeric}
\alias{eda.numeric}
\title{Generate EDA objects from numeric vectors.}
\usage{
\method{eda}{numeric}(x)
}
\arguments{
\item{x}{an object to be converted to EDA.}
}
\value{
an EDA object.
}
\description{
Generates EDA objects from a numeric vector.
}
