% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{eda}
\alias{eda}
\title{Generation of EDA objects.}
\usage{
eda(x, ...)
}
\arguments{
\item{x}{an object to be converted to EDA.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of class 'eda'.
}
\description{
Creates a eda object from multiple possible outputs such as vectors or data frames.
}
