% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{eda_classic_theme}
\alias{eda_classic_theme}
\alias{eda_void_theme}
\title{Custom \code{ggplot} themes.}
\usage{
eda_classic_theme(font_size = 8)

eda_void_theme(font_size = 8)
}
\arguments{
\item{font_size}{font size used in the theme. The default is 8.}
}
\value{
\code{ggplot}'s \code{theme} objects.
}
\description{
Themes compatible with requirements of most of scientific journals.
}
