% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_testing.R
\name{eff_size}
\alias{eff_size}
\title{Effect size for difference between EDA objects.}
\usage{
eff_size(
  ...,
  type = c("cohen_d", "paired_cohen_d", "wilcoxon_r", "paired_wilcoxon_r",
    "cohen_kappa", "freeman_theta", "cramer_v", "cohen_kappa", "etasq", "partial_etasq",
    "kruskal_etasq", "kendall_w"),
  exact = TRUE,
  ci = TRUE,
  boot_method = c("percentile", "bca", "normality")
)
}
\arguments{
\item{...}{EDA objects, at least two, created by \code{\link{eda}}.}

\item{type}{type of statistic the effect size. For two numeric EDA objects: independent variable Cohen's D
('cohens_d', \code{\link[rstatix]{cohens_d}}), paired Cohen's D ('paired_cohen_d', \code{\link[rstatix]{cohen_d}}),
independent variable Wilcoxon r ('wilcoxon_r', \code{\link[rstatix]{wilcox_effsize}}) or paired Wilcoxon r
('paired_wilcoxon_r', \code{\link[rstatix]{wilcox_effsize}}) are currently implemented. For two factor EDA objects:
Cohen's Kappa ('cohen_kappa', \code{\link[vcd]{Kappa}}) and Freeman's theta
('freeman_theta', \code{\link[rcompanion]{freemanTheta}}) is implemented. For 2 or more factor EDAs: Cramer's V
('cramer_v', \code{\link[DescTools]{CramerV}}) can be calculated. For two or more numeric EDAs: ANOVA-derived (partial)
eta-squared ('etasq' or 'partial_etasq', \code{\link[rstatix]{eta_squared}}) or
Kruskal-Wallis-derived eta-squared ('kruskal_etasq', \code{\link[rstatix]{kruskal_effsize}}) and
for aired data Kendall's W ('kendall_w', \code{\link[rstatix]{friedman_effsize}}) are currently available.}

\item{exact}{logical, should exact values for Chi-squared. Mann-Whitney and Wilcoxon test be returned?}

\item{ci}{logical, should confidence intervals for the test effect size be returned?}

\item{boot_method}{indicates how the bootstrap confidence intervals are calculated.
Can be any of 'percentile', 'bca', or 'normality', defaults to 'percentile'.}
}
\value{
an eTest object with the effect size statistic name as 'estimate' variable, effect size statistic value and
with 95/% confidence intervals.
}
\description{
compare two or more EDA objects with a range of statistical tests.
}
\details{
EDA object type is coerced to factor or numeric, as appropriate for the requested analysis. The default
statistic test results are returned as well: T test for Cohen's D, Mann-Whitney/Wilcoxon test for Wilcoxon r,
Chi-squared test results for Cohen's kappa, Cramer's V and Freeman's theta, one-way ANOVA for eta-squared and
partial eta-squared, Kruskal-Wallis test for Kruskal eta-squared and Friedman test for Kendall#s W. The p value
referst to the statistical test result.
}
