% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{etest}
\alias{etest}
\alias{is_etest}
\title{Generate \code{eTest} objects.}
\usage{
etest(
  test,
  stat_name = NA,
  stat = NA,
  df1 = NA,
  df2 = NA,
  estimate_name = NA,
  estimate = NA,
  lower_ci = NA,
  upper_ci = NA,
  p_value = NA,
  n = NA
)

is_etest(x)
}
\arguments{
\item{test}{name of the statistical test.}

\item{stat_name}{name of the test statistic.}

\item{stat}{value of the test statistic.}

\item{df1}{= degrees of freedom.}

\item{df2}{degrees of freedom.}

\item{estimate_name}{name of the test estimate.}

\item{estimate}{value of the test estimate.}

\item{lower_ci}{value of the lower confidence interval limit.}

\item{upper_ci}{value of the upper confidence interval limit.}

\item{p_value}{p value.}

\item{n}{number of complete observations used for testing.}

\item{x}{an object.}
}
\value{
an \code{etest} data frame. It shares most of its methods
with a "canonical" data frame.
}
\description{
Generates \code{eTest} objects storing results of statistical hypothesis tests
in a standardized data frame.
}
