% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{factor}
\alias{factor}
\alias{factor.default}
\alias{as.numeric}
\alias{as.numeric.default}
\alias{as.integer}
\alias{as.integer.default}
\title{Convert between types.}
\usage{
factor(x, ...)

\method{factor}{default}(x, ...)

as.numeric(x, ...)

\method{as.numeric}{default}(x, ...)

as.integer(x, ...)

\method{as.integer}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{additional arguments passed to methods.}
}
\value{
an object of the requested data type.
}
\description{
Generic functions for conversion between data types.
}
\details{
See: \code{\link[base]{factor}}, \code{\link[base]{as.numeric}}, and
\code{\link[base]{as.integer}}.
}
