% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_effect}
\alias{find_effect}
\title{Statistical significance and regulation sign.}
\usage{
find_effect(x, p_variable = c("p_adjusted", "p_value"))
}
\arguments{
\item{x}{a data frame with results of statistical hypothesis testing.}

\item{p_variable}{name of the variable storing p values to infer the
significance information.
If \code{p_variable = "p_value"}, raw p values will be used.}
}
\value{
the input data frame appended with columns \code{significant} and
\code{regulation} which store information on significance and sign of the effect.
}
\description{
The function infers statistical significance and sign of the effect
from results of statistical hypothesis testing.
Intended for internal use.
}
