% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_p}
\alias{format_p}
\title{Format p values in results of statistical hypothesis tests.}
\usage{
format_p(x, significant_p = 0.05, simplify_p = 0.001, signif_digits = 2)
}
\arguments{
\item{x}{a numeric vector.}

\item{significant_p}{significance threshold, defaults to p = 0.05.}

\item{simplify_p}{a numeric value, p values below \code{simplify_p} are presented
in a form \code{p < ...}, e.g. for the default p values < 0.001 be presented in a
p < 0.001 form.}

\item{signif_digits}{number of significant digits used for rounding of
p values.}
}
\value{
a character vector with the formatted p values.
}
\description{
Formatting of p values.
P values greater or equal than 0.05 are presented in a \code{ns (p = ...)} form.
P values lower that \code{simplify_p} value are presented in a \code{p < ...} form.
}
