% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{frequency.eda}
\alias{frequency.eda}
\title{Count observations categorical EDA objects to a table.}
\usage{
\method{frequency}{eda}(x, .drop = FALSE)
}
\arguments{
\item{x}{an EDA object, created by \code{\link{eda}}.}

\item{.drop}{logical, should empty levels be dropped?}
}
\value{
a tibble with the columns 'category', 'n', 'fraction' and 'percent'.
}
\description{
counts of observations assigned to each category or occurrence of unique values
for numeric EDAs.
}
\details{
NAs are listed as a separate category and empty levels are not skipped by default.
}
