% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{is_etest}
\alias{is_etest}
\title{Test the eTest class.}
\usage{
is_etest(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if the object is an instance of the eTest class.
}
