% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jitter_coordinates}
\alias{jitter_coordinates}
\title{Move plotting coordinates by random values.}
\usage{
jitter_coordinates(x, x_variable, x_sd = NULL, y_variable, y_sd = NULL)
}
\arguments{
\item{x}{a data frame.}

\item{x_variable}{name of the plotting variable to be presented in
the X axis.}

\item{x_sd}{\code{NULL} or value of standard deviation for the \code{x_variable}
modification.}

\item{y_variable}{name of the plotting variable to be presented in
the Y axis.}

\item{y_sd}{\code{NULL} or value of standard deviation for the \code{y_variable}
modification.}
}
\value{
the input data frame appended with columns \code{.x} and \code{.y} with
the modified plotting coordinates.
}
\description{
The function is intended for internal use.
Plotting coordinates specified by two numeric variables are modified
by addition of a random value drawn from normal distribution (mean is zero,
the standard deviation provided by the user).
}
\details{
When standard deviation values are \code{NULL}, they will be set to 10\% of
actual standard deviation values of \code{x_variable} and \code{y_variable}.
}
