% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_methods.R
\name{kurtosis.default}
\alias{kurtosis.default}
\title{Default kurtosis calculation method.}
\usage{
\method{kurtosis}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a numeric.
}
\description{
A wrapper around \code{\link[moments]{kurtosis}}
}
