% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate kurtosis.}
\usage{
kurtosis(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a numeric.
}
\description{
Calculates kurtosis of an object.
}
