% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_methods.R
\name{mean.eda}
\alias{mean.eda}
\title{Mean of an EDA object.}
\usage{
\method{mean}{eda}(eda_object, plain = FALSE)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}

\item{plain}{logical, should the output be coerced to a single vector?}
}
\value{
a tibble or a numeric value.
}
\description{
Gets arithmetic mean of a numeric-type EDA object. For factor-type ones
NULL is returned and a warning generated. NAs are removed before computation.
}
