% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot_utils.R
\name{multiplot}
\alias{multiplot}
\title{Plot two or more EDA objects.}
\usage{
multiplot(
  ...,
  eda_names = NULL,
  type = c("default", "bar", "violin", "box", "hist", "correlation", "paired", "bubble",
    "stack"),
  scale = c("none", "fraction", "percent"),
  point_alpha = 0.5,
  point_hjitter = 0.05,
  point_wjitter = 0.1,
  point_color = "steelblue",
  point_size = 2,
  line_color = "black",
  line_alpha = 0.25,
  cust_theme = ggplot2::theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  show_trend = TRUE,
  show_labels = TRUE,
  signif_digits = 2,
  txt_size = 2.75,
  txt_color = "black",
  geom_label = TRUE,
  bins = NULL,
  facet_hist = c("none", "horizontal", "vertical")
)
}
\arguments{
\item{...}{numeric-type EDA objects, at least two, created by \code{\link{eda}}.}

\item{eda_names}{a vector with names of the EDA objects.}

\item{type}{type of the plot. 'default' plots violin for numeric EDAs and bars for factors.
'bar', 'bubble' or 'stack' (stack-bar plot) are available for factor-type EDAs.
'violin', 'box', 'hist', 'correlation' and 'paired'
are available for numeric-type objects.}

\item{scale}{the feature to be presented in factor bar or bubble plots.
'none' plots counts, 'percent' plots percentages, 'fraction' presents fraction fo complete observations.}

\item{point_alpha}{alpha of the plot points.}

\item{point_hjitter}{point jitter height.}

\item{point_wjitter}{point jitter width.}

\item{point_color}{color of the points in the correlation plot.}

\item{point_size}{size of the points in the plots.}

\item{line_color}{color of the trend line in the correlation plots or the connecting lines in the paired plots.}

\item{line_alpha}{opacity of the connecting lines in the paired plot.}

\item{cust_theme}{custom ggplot2 theme.}

\item{plot_title}{text to be presented in the plot title.}

\item{plot_subtitle}{text to be presented in the plot subtitle.}

\item{x_lab}{text to be presented in the X axis title.}

\item{y_lab}{text to be presented in the Y axis title.}

\item{show_trend}{logical, should a trend line with 95\% confidence intervals be presented in the correlation plots?}

\item{show_labels}{logical, should labels with count numbers, percentages or fractions be presented in bar plots?}

\item{signif_digits}{significant digits used for the label value rounding.}

\item{txt_size}{size of the text label.}

\item{txt_color}{color of the text label.}

\item{geom_label}{logical, should the text in the stacked bar plot be
presented as a ggplot's geom_label?}

\item{bins}{bin number, passed to \code{\link[ggplot2]{histogram}}.}

\item{facet_hist}{'none': histograms are overlaid, 'horizontal': horizontal or 'vertical': vertical faceting.}
}
\description{
Plots values of two or more EDA objects as a classical bar, violin,
boxplot or correlation point plot representation.
}
\details{
the particular EDA objects are color coded.
}
