% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{na.omit.eda}
\alias{na.omit.eda}
\alias{na.exclude.eda}
\title{\code{NA} removal from \code{eda} objects.}
\usage{
\method{na.omit}{eda}(object, ...)

\method{na.exclude}{eda}(object, ...)
}
\arguments{
\item{object}{an \code{\link{eda}} object.}

\item{...}{arguments for methods.}
}
\value{
an \code{eda} object without \code{NA} values.
}
\description{
removes \code{NA} values from the \code{eda} objects.
}
