% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_general_methods.R
\name{nobs.eda}
\alias{nobs.eda}
\title{Number of (complete) observations in an \code{eda} object.}
\usage{
\method{nobs}{eda}(object, plain = FALSE, ...)
}
\arguments{
\item{object}{an \code{\link{eda}} object.}

\item{plain}{logical, should the values be returned as a numeric vector.}

\item{...}{additional arguments passed to methods, currently none.}
}
\value{
a tibble or numeric vector with the number of all and complete
observations.
}
\description{
counts the number of all (complete if na.rm is set to TRUE)
observations in the EDA object.
}
