% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corelation.R
\name{normality}
\alias{normality}
\title{Test normality of an EDA object.}
\usage{
normality(eda_object)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}
}
\value{
an eTest object with the W statistic and p value.
}
\description{
tests normality of an EDA object values with Shapiro-Wilk test. For factor-type ones
NULL is returned and a warning generated.
}
\details{
a wrapper around \code{\link[stats]{shapiro.test}}.
}
