% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot.eda}
\alias{plot.eda}
\title{Plot method for EDA objects}
\usage{
\method{plot}{eda}(
  x,
  type = c("default", "bar", "violin", "box", "hist", "qq"),
  point_color = "steelblue",
  point_alpha = 0.5,
  point_hjitter = 0.05,
  point_wjitter = 0.1,
  fill_color = "steelblue",
  line_color = "orangered3",
  txt_size = 2.75,
  signif_digits = 2,
  cust_theme = ggplot2::theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  ...
)
}
\arguments{
\item{x}{eda object, created by \code{\link{eda}}.}

\item{type}{plot type. 'default' is bar for factors and violin for numerics. 'bar' is available for factor-type EDAs.
'violin', 'box', 'hist' and 'qq' are available for numeric-type objects.}

\item{point_color}{color of the plot points.}

\item{point_alpha}{alpha of the plot points.}

\item{point_hjitter}{point jitter height.}

\item{point_wjitter}{point jitter width.}

\item{fill_color}{color of the bars and violins.}

\item{line_color}{color of the qqline in the qq plot.}

\item{txt_size}{size of the plot label text.}

\item{signif_digits}{significant digits used for rounding of the statistic summary.}

\item{cust_theme}{custom ggplot2 theme.}

\item{plot_title}{text to be presented in the plot title.}

\item{plot_subtitle}{text to be presented in the plot subtitle.}

\item{...}{extra arguments passed to \code{\link[ggplot]{geom_violin}}, \code{\link[ggplot]{geom_histogram}} or
\code{\link[ggplot]{geom_boxplot}}.}
}
\value{
a ggplot.
}
\description{
Plots distribution of the given EDA object.
}
\details{
the summary of EDA is presented in the plot tag. The type conversion is done prior to plotting.
}
