% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etest_methods.R
\name{plot.etest}
\alias{plot.etest}
\title{Plotting of effect size and significance for \code{etest} class object.}
\usage{
\method{plot}{etest}(
  x,
  p_variable = c("p_adjusted", "p_value"),
  show_txt = TRUE,
  txt_size = 2.75,
  n_top = 10,
  label_type = c("text", "label"),
  labeller_fun = identity,
  palette = etest_colors(),
  fill_lab = "",
  point_size = 2,
  point_alpha = 1,
  point_wjitter = 0,
  point_hjitter = 0,
  show_lines = TRUE,
  line_type = "dashed",
  line_width = 0.75,
  line_alpha = 1,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  cust_theme = eda_classic_theme(),
  ...
)
}
\arguments{
\item{x}{an object of \code{\link{etest}} class.}

\item{p_variable}{name of the variable storing p values to be presented in the
plot. If \code{p_variable = "p_value"}, raw p values will be presented.}

\item{show_txt}{logical, should text labels with frequencies be
displayed in the plot?}

\item{txt_size}{size of text labels in the plots.}

\item{n_top}{number of top effects (absolute value of effect size) to be
labeled with variable names in the plot.}

\item{label_type}{type of the labels of the top strongest effects: text
generated with \code{\link[ggrepel]{geom_text_repel}} or labels created with
\code{\link[ggrepel]{geom_label_repel}}.}

\item{labeller_fun}{a function used to transform names of variables
presented in data point labels.}

\item{palette}{a vector with color names or hex codes for fill scales.}

\item{fill_lab}{title for the fill scale.}

\item{point_size}{size of data points}

\item{point_alpha}{alpha (opacity) of data points.}

\item{point_wjitter}{width of jittering of data points. If \code{NULL}, jittering
will be determined with a simple heuristics
(10\% of standard deviation of the plotting variable).}

\item{point_hjitter}{height of jittering of data points. If \code{NULL}, jittering
will be determined with a simple heuristics
(10\% of standard deviation of the plotting variable).}

\item{show_lines}{logical: should lines representing cutoffs of statistical
significance and null effect size be displayed in the plots?
\code{TRUE} by default.}

\item{line_type}{type of the lines representing the significance
and effect size cutoffs. Dashed by default.}

\item{line_width}{width of interquartile whiskers in violin plots, and
lines depicting statistics in histograms and QQ plots.}

\item{line_alpha}{alpha/opacity of the lines representing the significance
and effect size cutoffs. Dashed by default.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle. If \code{NULL}, range of N numbers of observations
used in the statistical test will appear there.}

\item{x_lab}{title of the X axis.}

\item{y_lab}{title of the Y axis.}

\item{cust_theme}{custom \code{ggplot}'s \code{theme} object or \code{NULL}. If \code{NULL} no
custom theme is applied, which makes the plot objects much smaller.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_point}}.}
}
\value{
a list of \code{ggplot} objects.
}
\description{
The \code{plot()} method for \code{etest} data frames with results of statistical
hypothesis testing.
The function generates a list of scatter plots (one for each test type)
with effect size statistic values on the X axis and raw p values or p values
adjusted for multiple testing on the Y axis.
Single variables are represented by points; point color codes for statistical
significance, and optionally, regulation sign.
}
