% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_plotting_utilities.R
\name{plot_2df_numeric}
\alias{plot_2df_numeric}
\alias{plot_2df_factor}
\title{Scatter, heat map, and dot plots for two variables.}
\usage{
plot_2df_numeric(
  data,
  variable1,
  variable2,
  point_color = "steelblue",
  point_size = 2,
  point_alpha = 0.75,
  point_hjitter = 0,
  point_wjitter = 0,
  show_trend = TRUE,
  line_color = "gray40",
  line_width = 0.75,
  line_alpha = 0.5,
  cust_theme = eda_classic_theme(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  ...
)

plot_2df_factor(
  data,
  variable1,
  variable2,
  type = c("heat_map", "bubble"),
  scale = c("none", "percent"),
  shape_color = "black",
  shape_alpha = 1,
  show_txt = TRUE,
  txt_size = 2.75,
  txt_vjust = NULL,
  txt_hjust = NULL,
  txt_x_nudge = 0.1,
  txt_style = c("minimal", "full"),
  signif_digits = 2,
  cust_theme = eda_classic_theme(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  fill_lab = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variable1}{name of the data frame's variable whose distribution will be
shown in the X axis of the plot. Provided as a string.}

\item{variable2}{name of the data frame's variable whose distribution will be
shown in the Y axis of the plot. Provided as a string.}

\item{point_color}{color of data points.}

\item{point_size}{size of data points}

\item{point_alpha}{alpha (opacity) of data points.}

\item{point_hjitter}{horizontal jittering of data points}

\item{point_wjitter}{vertical jittering of data points.}

\item{show_trend}{logical, should the trend line be displayed in the plot?
The line is generated with \code{\link[ggplot2]{geom_smooth}}.}

\item{line_color}{color of the trend line.}

\item{line_width}{width of the trend line.}

\item{line_alpha}{alpha (opacity) of the trend line.}

\item{cust_theme}{custom \code{ggplot}'s \code{theme} object or \code{NULL}. If \code{NULL} no
custom theme is applied, which makes the plot objects much smaller.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle; if \code{NULL} numbers of all and complete
observations are shown here.}

\item{x_lab}{title of the X axis.}

\item{y_lab}{title of the Y axis, if \code{NULL} the title is set with a simple
heuristics (e.g. if \code{scale = "none"}, \verb{"observations, N}).}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_smooth}},
\code{\link[ggplot2]{geom_tile}}, or \code{\link[ggplot2]{geom_point}}.}

\item{type}{plot type.}

\item{scale}{type of frequency data to be presented in bar and stack plots:
none (counts) or percents,}

\item{shape_color}{color of rims of bars, boxes, violins, and histograms.}

\item{shape_alpha}{alpha (opacity) of bars, boxes, violins, and histograms.}

\item{show_txt}{logical, should text labels with frequencies be
displayed in the plot?}

\item{txt_size}{size of text labels in the plots.}

\item{txt_vjust}{vertical justification of the text labels in the plots.
Plot type-specific defaults are applied if set to \code{NULL}.}

\item{txt_hjust}{horizontal justification of the text labels in the plots.
Plot type-specific defaults are applied if set to \code{NULL}.}

\item{txt_x_nudge}{X axis nudging/offset of text labels with frequencies in
bubble plots.}

\item{txt_style}{style of the frequencies displayed in the plot.
If \code{txt_style = "minimal"} (default), numbers of observations or percentages
are displayed in the plot, as appropriate for \code{scale}.
If \code{txt_style = "full"}, both percentages and numbers of observations are
shown.}

\item{signif_digits}{significant digits used to round numeric values displayed
in the plot (e.g. percentages).}

\item{fill_lab}{title of the fill scale.}
}
\value{
a \code{ggplot} object.
}
\description{
Plots for two variables in a data frame.
For a pair of numeric variables, a scatter plot with an optional trend/standard
error line is implemented.
For a pair of factors, a heat map or a bubble plot with counts or percentages
of complete observations is available.
}
\details{
The observations presented in the plot are restricted to cases with complete
\code{variable1} and variable \code{variable2} information.
}
