% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_bar}
\alias{plot_bar}
\title{Bar plot.}
\usage{
plot_bar(
  eda_object,
  signif_digits = 2,
  fill_color = "steelblue",
  cust_theme = ggplot2::theme_classic(),
  txt_size = 2.75
)
}
\arguments{
\item{eda_object}{an EDA object.}
}
\value{
a ggplot object.
}
\description{
creates a bar plot for a factor-type EDA object.
}
