% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exploration.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot correlation of two variables.}
\usage{
plot_correlation(
  data,
  variables,
  type = c("correlation", "bar", "bubble"),
  scale = c("none", "percent", "fraction"),
  point_alpha = 0.5,
  point_hjitter = 0.05,
  point_wjitter = 0.1,
  point_color = "steelblue",
  point_size = 2,
  line_color = "black",
  line_alpha = 0.25,
  cust_theme = ggplot2::theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  show_trend = TRUE,
  show_labels = TRUE,
  signif_digits = 2,
  txt_size = 2.75
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector with variable names. If more than two provided, only the first two are analyzed.}

\item{type}{type of the plot. 'correlation' returns a standard point plot, 'bar' returns a bar plot of
frequencies, 'bubble' represents the contingency matrix as a bubble plot.}

\item{scale}{the feature to be presented in factor bar plots. 'none' plots counts, 'percent' plots percentages,
'fraction' presents fraction fo complete observations.}

\item{point_alpha}{alpha of the plot points.}

\item{point_hjitter}{point jitter height.}

\item{point_wjitter}{point jitter width.}

\item{point_color}{color of the points in the correlation plot.}

\item{point_size}{size of the points in the plots.}

\item{line_color}{color of the trend line in the correlation plots or the connecting lines in the paired plots.}

\item{line_alpha}{opacity of the connecting lines in the paired plot.}

\item{cust_theme}{custom ggplot2 theme.}

\item{plot_title}{text to be presented in the plot title.}

\item{plot_subtitle}{text to be presented in the plot subtitle.}

\item{x_lab}{text to be presented in the X axis title.}

\item{y_lab}{text to be presented in the Y axis title.}

\item{show_trend}{logical, should a trend line with 95\% confidence intervals be presented in the correlation plots?}

\item{show_labels}{logical, should labels with count numbers, percentages or fractions be presented in bar plots?}

\item{signif_digits}{significant digits used for the label value rounding.}

\item{txt_size}{size of the text label.}
}
\description{
plots correlation of two variables. Point plot is available for numeric features
and bar plot or bubble plot of the contingency matrix for factors.
}
