% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_plotting_utilities.R
\name{plot_df_factor}
\alias{plot_df_factor}
\alias{plot_df_numeric}
\alias{plot_df_histogram}
\alias{plot_df_qq}
\title{Plotting utilities for factor and numeric types.}
\usage{
plot_df_factor(
  data,
  variable,
  split_factor,
  .drop = TRUE,
  type = c("stack", "bar", "bubble"),
  scale = c("none", "percent"),
  signif_digits = 2,
  shape_color = "black",
  shape_alpha = 1,
  dodge_width = 0.9,
  cust_theme = theme_classic(),
  show_txt = TRUE,
  txt_size = 2.75,
  txt_vjust = NULL,
  txt_hjust = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  fill_lab = NULL,
  x_n_labs = TRUE,
  labeller_fun = identity,
  ...
)

plot_df_numeric(
  data,
  variable,
  split_factor,
  type = c("violin", "box", "paired"),
  show_stats = TRUE,
  shape_color = "black",
  shape_alpha = 0.3,
  point_color = "gray40",
  point_size = 2,
  point_alpha = 0.75,
  point_hjitter = 0,
  point_wjitter = 0.1,
  line_color = NULL,
  line_width = 0.75,
  cust_theme = theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  fill_lab = NULL,
  x_n_labs = TRUE,
  labeller_fun = identity,
  ...
)

plot_df_histogram(
  data,
  variable,
  split_factor,
  type = c("histogram", "density"),
  show_stats = TRUE,
  shape_color = "black",
  shape_alpha = 0.3,
  line_color = "orangered3",
  line_width = 0.75,
  cust_theme = theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  fill_lab = NULL,
  x_n_labs = TRUE,
  labeller_fun = identity,
  facet = c("none", "horizontal", "vertical"),
  facet_scales = "free",
  facet_space = "free",
  ...
)

plot_df_qq(
  data,
  variable,
  split_factor,
  point_size = 2,
  point_alpha = 0.75,
  point_hjitter = 0,
  point_wjitter = 0,
  line_width = 0.75,
  cust_theme = theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "quantiles, theoretical normal distribution",
  y_lab = "quantiles, observed distribution",
  fill_lab = NULL,
  x_n_labs = TRUE,
  labeller_fun = identity,
  facet = c("none", "horizontal", "vertical"),
  facet_scales = "free",
  facet_space = "free",
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variable}{name of the data frame's variable whose distribution will be
shown in the plot. Provided as a string.}

\item{split_factor}{name of the data frame's variable, whose categories will
be used to split values of \code{variable} in the plot. Provided as a string.}

\item{.drop}{should empty levels of a factor be dropped.}

\item{type}{plot type.}

\item{scale}{type of frequency data to be presented in bar and stack plots:
none (counts) or percents,}

\item{signif_digits}{significant digits used to round numeric values displayed
in the plot (e.g. percentages).}

\item{shape_color}{color of rims of bars, boxes, violins, and histograms.}

\item{shape_alpha}{alpha (opacity) of bars, boxes, violins, and histograms.}

\item{dodge_width}{spacing between bars in bar plots.}

\item{cust_theme}{custom \code{ggplot}'s \code{theme} object.}

\item{show_txt}{logical, should text labels with frequencies be
displayed in the plot?}

\item{txt_size}{size of text labels in the plots.}

\item{txt_vjust}{vertical justification of the text labels in the plots.
Plot type-specific defaults are applied if set to \code{NULL}.}

\item{txt_hjust}{horizontal justification of the text labels in the plots.
Plot type-specific defaults are applied if set to \code{NULL}.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle; if \code{NULL} numbers of all and complete
observations are shown here.}

\item{x_lab}{title of the X axis.}

\item{y_lab}{title of the Y axis, if \code{NULL} the title is set with a simple
heuristics (e.g. if \code{scale = "none"}, \verb{"observations, N}).}

\item{fill_lab}{title of the fill scale.}

\item{x_n_labs}{logical. If \code{TRUE}, N numbers of complete observations in
strata of \code{variable} defined by the \code{split_factor}'s categories are displayed
in the X axis labels. Defaults to \code{TRUE}.}

\item{labeller_fun}{a function used to transform names of \code{split_factor}'s
categories presented in the plot axes. Defaults to \code{identity}, which means
that the category names are displayed as they are specified by levels of the
splitting factor.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_boxplot}},
\code{\link[ggplot2]{geom_violin}}, \code{\link[ggplot2]{geom_histogram}}.}

\item{show_stats}{logical. Histograms: should vertical lines representing
median (solid) and interquartile range (dashed) be shown?
Violin plots: should diamonds and whiskers representing median with
interquartile range be displayed?}

\item{point_color}{color of data points.}

\item{point_size}{size of data points}

\item{point_alpha}{alpha (opacity) of data points.}

\item{point_hjitter}{horizontal jittering of data points}

\item{point_wjitter}{vertical jittering of data points.}

\item{line_color}{color of interquartile whiskers in violin plots, and
lines depicting statistics in histograms.}

\item{line_width}{width of interquartile whiskers in violin plots, and
lines depicting statistics in histograms.}

\item{facet}{"none": histograms/density plots are overlaid,
"horizontal": horizontal or "vertical": vertical faceting.}

\item{facet_scales}{specifies scales in the faceted histogram/density plot,
see: \code{\link[ggplot2]{facet_grid}} for details.}

\item{facet_space}{specifies space in the faceted histogram/density plot,
see: \code{\link[ggplot2]{facet_grid}} for details.}
}
\value{
a \code{ggplot} object.
}
\description{
Internal functions which generate \code{ggplot} plots for data frame's variables
of different data types.
For factors, stack, bar, and bubble plots are available with counts or
percentages of observations in categories of a splitting factor.
For numeric variables, violin and box plots, histograms, density plots,
and quantile - quantile (QQ) plots are implemented.
}
\details{
If \code{type = "paired"}, a before-after plot is generated and its assumed that
the plotting data are already arranged by the blocking or identified variable.
}
