% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_plotting_utilities.R
\name{plot_factor}
\alias{plot_factor}
\alias{plot_numeric}
\alias{plot_histogram}
\alias{plot_qq}
\title{Plot utilities for single \code{eda} objects.}
\usage{
plot_factor(
  eda_object,
  .drop = TRUE,
  type = c("stack", "bar", "bubble"),
  scale = c("none", "percent"),
  signif_digits = 2,
  palette = tableau10_colors(),
  shape_fill = "steelblue",
  shape_color = "black",
  shape_alpha = 1,
  cust_theme = eda_classic_theme(),
  show_txt = TRUE,
  txt_size = 2.75,
  txt_vjust = NULL,
  txt_hjust = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_txt = "",
  x_lab = "category",
  ...
)

plot_numeric(
  eda_object,
  type = c("violin", "box"),
  show_stats = TRUE,
  shape_fill = "steelblue",
  shape_color = "black",
  shape_alpha = 0.3,
  point_color = "gray40",
  point_size = 2,
  point_alpha = 0.75,
  point_hjitter = 0,
  point_wjitter = 0.1,
  line_color = "orangered3",
  line_width = 0.75,
  cust_theme = eda_classic_theme(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_txt = "",
  x_lab = "",
  y_lab = "",
  ...
)

plot_histogram(
  eda_object,
  type = c("histogram", "density"),
  shape_fill = "steelblue",
  shape_color = "black",
  shape_alpha = 0.75,
  show_stats = TRUE,
  line_color = "orangered3",
  line_width = 0.75,
  cust_theme = eda_classic_theme(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "value",
  y_lab = NULL,
  ...
)

plot_qq(
  eda_object,
  point_color = "steelblue",
  point_size = 2,
  point_alpha = 0.75,
  point_hjitter = 0,
  point_wjitter = 0,
  line_color = "orangered3",
  line_width = 0.75,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "quantiles, theoretical normal distribution",
  y_lab = "quantiles, observed distribution",
  cust_theme = eda_classic_theme(),
  ...
)
}
\arguments{
\item{eda_object}{and \code{\link{eda}} object.}

\item{.drop}{logical, should empty levels of a factor be dropped?}

\item{type}{plot type.}

\item{scale}{type of frequency data to be presented in bar and stack plots:
none (counts) or percents,}

\item{signif_digits}{significant digits used to round numeric values displayed
in the plot (e.g. percentages).}

\item{palette}{a vector with color names or hex codes for fill scales.}

\item{shape_fill}{fill color of bars, boxes, violins, and histograms.}

\item{shape_color}{color of rims of bars, boxes, violins, and histograms.}

\item{shape_alpha}{alpha (opacity) of bars, boxes, violins, and histograms.}

\item{cust_theme}{custom \code{ggplot}'s \code{theme} object or \code{NULL}. If \code{NULL} no
custom theme is applied, which makes the plot objects much smaller.}

\item{show_txt}{logical, should text labels with frequencies be
displayed in the plot?}

\item{txt_size}{size of text labels in the plots.}

\item{txt_vjust}{vertical justification of the text labels in the plots.
Plot type-specific defaults are applied if set to \code{NULL}.}

\item{txt_hjust}{horizontal justification of the text labels in the plots.
Plot type-specific defaults are applied if set to \code{NULL}.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle; if \code{NULL} numbers of all and complete
observations are shown here.}

\item{x_txt}{text to be displayed in the X axis. Concerns stack, box,
and violin plots.}

\item{x_lab}{title of the X axis.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_boxplot}},
\code{\link[ggplot2]{geom_violin}}, \code{\link[ggplot2]{geom_histogram}},
and \code{\link[ggplot2]{geom_density}}.}

\item{show_stats}{logical. Histograms: should vertical lines representing
median (solid) and interquartile range (dashed) be shown?
Violin plots: should diamonds and whiskers representing median with
interquartile range be displayed?}

\item{point_color}{color of data points.}

\item{point_size}{size of data points}

\item{point_alpha}{alpha (opacity) of data points.}

\item{point_hjitter}{horizontal jittering of data points}

\item{point_wjitter}{vertical jittering of data points.}

\item{line_color}{color of interquartile whiskers in violin plots, and
lines depicting statistics in histograms and QQ plots.}

\item{line_width}{width of interquartile whiskers in violin plots, and
lines depicting statistics in histograms and QQ plots.}

\item{y_lab}{title of the Y axis.}
}
\value{
A \code{ggplot} object.
}
\description{
Bar, stack, bubble, violin and box plots, histograms, density plots, violin and
quantile - quantile (QQ) plots for factor and numeric
\code{\link{eda}} objects.
}
