% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histogram.}
\usage{
plot_histogram(
  eda_object,
  signif_digits = 2,
  fill_color = "steelblue",
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{eda_object}{an EDA object.}
}
\value{
a ggplot object.
}
\description{
creates a histogram for a numeric-type EDA object.
}
