% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_multi_variables}
\alias{plot_multi_variables}
\title{Plot distribution of multiple variables in a data frame.}
\usage{
plot_multi_variables(data, variables, split_factor = NULL, type = NULL, ...)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a character vector with variable names.}

\item{split_factor}{optional, a character string with the name of the factor
defining analysis groups. If \code{split_factor = NULL}, the distributions in
the whole data frame will be shown.}

\item{type}{plot type. For numeric variables: "violin", "box", "histogram",
"density", "bar", "ribbon", and "forest". For factors: "stack".
If \code{type = NULL} (default), a panel of violin plots is generated for numeric
variables, and a panel of stack plots is returned for factor variables.}

\item{...}{additional arguments passed to internal plotting functions
(for violin. box, bar, ribbon, and Forest plots: \code{\link{draw_numeric_panel}};
for stack plots of factor frequencies: \code{\link{draw_factor_panel}}).
They specify, among others, type of frequency data presented in the plots,
appearance of shapes and data point, and text labels.}
}
\value{
a \code{ggplot} object.
}
\description{
Function \code{plot_multi_variables()} generates compound plots or panels
which display distribution of multiple numeric or factor variables in a
data frame.
}
\details{
Variables specified by \code{variable} character vector of names have to be
all numeric or all factor objects - mixed types cause an error.
}
