% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_two_variables}
\alias{plot_two_variables}
\alias{plot_correlation}
\title{Plot distribution of two variables in a data frame.}
\usage{
plot_two_variables(data, variable1, variable2, type = NULL, ...)

plot_correlation(data, variables, type = NULL, ...)
}
\arguments{
\item{data}{a data frame.}

\item{variable1}{name of the data frame's variable whose distribution will be
shown in the X axis of the plot. Provided as a string.}

\item{variable2}{name of the data frame's variable whose distribution will be
shown in the Y axis of the plot. Provided as a string.}

\item{type}{plot type: "scatter" for a pair of numeric variables, "heat_map"
or "bubble" for a pair of factors. If \code{NULL}, a scatter plot is generated
for numeric, and a heat map is plotted for factor variables.}

\item{...}{additional arguments passed to internal plotting functions
(numeric variables: \code{\link{plot_2df_numeric}},
factor variables: \code{\link{plot_2df_factor}}).
They specify, among others, type of plotted frequency data, colors and opacity of
points and shapes, point sizes, text labels, and \code{ggplot} themes.}

\item{variables}{a character vector with names of the two variables to be
presented in the plot.}
}
\value{
a \code{ggplot} object.
}
\description{
Functions \code{plot_two_variables()} and \code{plot_correlation()} generate
\code{ggplot} plots of distribution of two variables.
For a pair of two numeric variables, a scatter plot is generated.
Frequencies of observations (counts or percentages of complete cases)
for a pair of factors are displayed in a heat map or a bubble plot.
}
\details{
The functions work for two numeric or two factor variables; if the variables'
types are mixed, and error is thrown.
In such cases, the user is better served by \code{\link{plot_variable}} function.
Function \code{plot_correlation()} is kept only for compatibility with previous
versions of the package and will be removed in future.
Please use \code{plot_two_variables()} instead,
}
