% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{relevel.eda}
\alias{relevel.eda}
\title{Relevel an EDA object.}
\usage{
\method{relevel}{eda}(eda_object, newlevels)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}

\item{newlevels}{a character vector with new levels.}
}
\value{
a factor-type EDA.
}
\description{
Sets new levels for a factor-type EDA object.
Releveling of a numeric-type EDA raises an error.
}
