% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{result_summary}
\alias{result_summary}
\title{Summary table with descriptive statistics and results of statistical hypothesis tests.}
\usage{
result_summary(
  x,
  y = NULL,
  labeller_fun = identity,
  test_columns = c("test", "raw_significance", "significance", "effect_size_txt"),
  ...
)
}
\arguments{
\item{x}{a data frame with descriptive statistics: an object of class
\code{\link{destat}} created with \code{\link{explore.data.frame}}.}

\item{y}{\code{NULL} or a data frame of \code{\link{etest}} with results of
statistical hypothesis testing created with \code{\link{compare_variables}}.}

\item{labeller_fun}{a function used to transform names of variables in the
result summary data frame.}

\item{test_columns}{a character vector of names of columns in \code{y} to be merged
with the descriptive statistic information.}

\item{...}{additional arguments, currently none.}
}
\value{
A data frame with formatted descriptive statistics and, optionally,
test type information, raw and adjusted p values, and effect sizes.
}
\description{
Function \code{result_summary()} creates a ready-to-use summary table which bundles
a data frame with descriptive statistics (created with \code{\link{explore.data.frame}})
with results of statistical hypothesis tests for comparison of expected values
(optional, created with \code{\link{compare_variables}}).
}
