% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_numeric_methods.R
\name{scale.eda}
\alias{scale.eda}
\title{Scaling and centering \code{eda} objects.}
\usage{
\method{scale}{eda}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{an \code{\link{eda}} object}

\item{center}{either a logical value or numeric-alike vector of length
equal to the \code{\link{eda}} object.
See \code{\link[base]{scale}} for details.}

\item{scale}{either a logical value or numeric-alike vector of length
equal to the \code{\link{eda}} object.
See \code{\link[base]{scale}} for details.}
}
\value{
a scaled, centered numeric \code{\link{eda}} object
}
\description{
Scaling and centering of an \code{eda} object as described
in \code{\link[base]{scale}}.
For factor objects, the unmodified object is returned with a warning.
}
\details{
Unlike the generic \code{\link[base]{scale}} function, it returns an
\code{\link{eda}} object.
Any \code{NA} values are silently removed.
}
