% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_methods.R
\name{sd.default}
\alias{sd.default}
\title{Default standard deviation calculation method.}
\usage{
\method{sd}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a numeric.
}
\description{
A wrapper around \code{\link[stats]{sd}}
}
