% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_methods.R
\name{sd.eda}
\alias{sd.eda}
\title{Standard deviance of an EDA object.}
\usage{
\method{sd}{eda}(x, plain = FALSE)
}
\arguments{
\item{x}{an EDA object, created by \code{\link{eda}}.}

\item{plain}{logical, should the output be coerced to a single vector?}
}
\value{
a tibble or a numeric value.
}
\description{
Gets standard deviation of a numeric-type EDA object. For factor-type ones
NULL is returned and a warning generated. NAs are removed before computation.
}
