% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sd}
\alias{sd}
\title{Calculate standard deviation.}
\usage{
sd(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a numeric.
}
\description{
Calculates standard deviation of an object.
}
