% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_methods.R
\name{skewness.default}
\alias{skewness.default}
\title{Default skewness calculation method.}
\usage{
\method{skewness}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a numeric.
}
\description{
A wrapper around \code{\link[moments]{skewness}}
}
