% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{skewness}
\alias{skewness}
\title{Calculate skewness.}
\usage{
skewness(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a numeric.
}
\description{
Calculates skewness of an object.
}
