% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_general_methods.R
\name{split.eda}
\alias{split.eda}
\title{Splitting of \code{eda} objects by levels of a factor.}
\usage{
\method{split}{eda}(x, f, drop = TRUE, ...)
}
\arguments{
\item{x}{an \code{\link{eda}} object.}

\item{f}{a factor or a factor \code{\link{eda}} object.}

\item{drop}{logical, should empty levels be dropped?}

\item{...}{extra arguments passed to methods.}
}
\value{
a list of \code{\link{eda}} objects.
}
\description{
Splitting of an \code{eda} object by levels of a factor (a plain factor or
an \code{eda} object).
}
\details{
The lengths of \code{x} and \code{f} vectors have to be equal.
If there are any \code{NA} values in the \code{f} vector, the corresponding values in
\code{x} are removed prior to splitting.#'
}
