% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.eda}
\alias{summary.eda}
\title{Summary for EDA objects.}
\usage{
\method{summary}{eda}(object, pub_styled = FALSE, signif_digits = 2)
}
\arguments{
\item{object}{eda object, created by \code{\link{eda}}.}

\item{pub_styled}{logical, should the output be publication-ready formatted? See the details.}

\item{signif_digits}{significant digits used for rounding in the publication-style output.}
}
\description{
Summary of numeric- and factor-type EDA objects.
}
\details{
The pub_style argument allows for generating a 'nicer' output accepted by multiple journals.
A tibble with the column 'statistic' is returned.
}
