% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.etest}
\alias{summary.etest}
\title{Summary for the eTest objects.}
\usage{
\method{summary}{etest}(etest_object, pub_styled = FALSE, signif_digits = 2, simplify_p = TRUE)
}
\arguments{
\item{etest_object}{an eTest object.}

\item{pub_styled}{logical, should the output be publication-ready formatted? See the details.}

\item{signif_digits}{significant digits used for rounding in the publication-style output.}

\item{simplify_p}{logical, should p_values < 0.001 be presented in a p < 0.001 form?}
}
\description{
Returns a tibble representation of the statistical testing for differences between EDA objects.
}
\details{
The pub_style argument allows for generating a 'nicer' output accepted by multiple journals.
A tibble with the columns 'test', 'test_stat', 'p_value', 'significance', 'effect size' and 'n' is returned.
}
