% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{tableau20_colors}
\alias{tableau20_colors}
\alias{tableau10_colors}
\alias{wes_grandbudapest1_colors}
\alias{wes_grandbudapest2_colors}
\alias{wes_zissou1_colors}
\alias{wes_fantasticfox1_colors}
\alias{wes_moonrise1_colors}
\title{Color palettes.}
\usage{
tableau20_colors(name = FALSE)

tableau10_colors(name = FALSE)

wes_grandbudapest1_colors(name = FALSE)

wes_grandbudapest2_colors(name = FALSE)

wes_zissou1_colors(name = FALSE)

wes_fantasticfox1_colors(name = FALSE)

wes_moonrise1_colors(name = FALSE)
}
\arguments{
\item{name}{logical should the vector be returned with the color names?}
}
\value{
named or unnamed character vectors with hex codes of colors.
}
\description{
A medley of functions returning named character vectors with
Tableau and Wes Anderson colors.
}
