% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_testing.R
\name{test}
\alias{test}
\title{Compare EDA objects.}
\usage{
test(
  ...,
  type = c("t_test", "paired_t_test", "sign_test", "mann_whitney_test",
    "wilcoxon_test", "anova", "rm_anova", "kruskal_test", "friedman_test", "chisq_test"),
  exact = TRUE,
  ci = TRUE,
  boot_method = c("percentile", "bca", "normality")
)
}
\arguments{
\item{...}{numeric-type EDA objects, at least two, created by \code{\link{eda}}.}

\item{type}{type of statistic test. For two numeric EDA objects: t test ('t_test', \code{\link[stats]{t.test}}),
paired t test ('paired_t_test', \code{\link[stats]{t.test}}), sign test ('sign_test', \code{\link[DescTools]{SignTest}}),
Mann-Whitney test ('mann_whitney_test', \code{\link[stats]{t.test}})
or paired Wilcoxon test ('wilcoxon_test', \code{\link[stats]{wilcox.test}}).
Multiple numeric-type EDA objects can be compared with one-way ANOVA ('anova', \code{\link[stats]{aov}}),
repeated measure one-way ANOVA ('rm_anova', \code{\link[stats]{aov}}), Kruskal-Wallis test
('kruskal_test', \code{\link[stats]{kruskal.test}}) or Friedman test for repeated measurements
('friedman_test', \code{\link[stats]{friedman.test}}). For factor-type EDA objects, Chi-squared test
may be applied ('chisq_test', \code{\link[stats]{chisq.test}}).}

\item{exact}{logical, should exact values for Chi-squared. Mann-Whitney and Wilcoxon test be returned?}

\item{ci}{logical, should confidence intervals for the test effect size be returned?}

\item{boot_method}{indicates how the bootstrap confidence intervals are calculated.
Can be any of 'percentile', 'bca', or 'normality', defaults to 'percentile'.}
}
\value{
an eTest object with the test name, test statistic value and, optionally,
the effect size with 95/% confidence intervals.
}
\description{
compare two or more EDA objects with a range of statistical tests.
}
\details{
EDA object type is coerced to factor or numeric, as appropriate for the requested analysis.
The default effect sizes returned as estimates with along with the test results include difference in mean for t tests,
difference in (pseudo-) median for sign, Mann-Whitney and Wilcoxon tests, eta-squared (% explained variance)
for one-way ANOVA and repeated measure one-way ANOVA, eta-square for Kruskal-Wallis test (calculated by
\code{\link[rstatix]{kruskal_effsize}}) and Kendall W for Friedman test (\code{\link[rstatix]{friedman_effsize}}).
For Chi-squared test, Cramer's V is returned (\code{\link[rcompanion]{cramerV}}).
}
