% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_tst_df}
\alias{validate_tst_df}
\title{Check validity of a data frame used in statistical tests.}
\usage{
validate_tst_df(data, variables = NULL, split_factor = NULL, coerce = TRUE)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a character vector of variables. If \code{NULL}, all variables of
the data frame except of \code{split_factor} are considered.}

\item{split_factor}{\code{NULL} or a name of a variable used as the splitting factor.}

\item{coerce}{logical: should character and logical variables be coerced to
factors with the default level setting option? Defaults to \code{TRUE}.}
}
\value{
a data frame with the variables and splitting factors of interest.
Variable formats (numeric or factor) are stored as \code{variable_format} attribute.
Empty categories of the splitting factor are automatically dropped.
}
\description{
An internal function testing if variables and splitting factors are present
in a data frame, and retrieve their formats.
}
