% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_methods.R
\name{var.default}
\alias{var.default}
\title{Default variance calculation method.}
\usage{
\method{var}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
a numeric.
}
\description{
A wrapper around \code{\link[stats]{var}}
}
