% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corelation.R
\name{variance}
\alias{variance}
\title{Compare variances of EDA objects.}
\usage{
variance(...)
}
\arguments{
\item{...}{numeric-type EDA objects, at least two, created by \code{\link{eda}}.}
}
\value{
an eTest object with the the test statistics and p values.
}
\description{
Compares variances of two or more numeric-type EDA objects with Levene test.
Raises an error is factor-type EDAs provided.
}
\details{
A wrapper around \code{\link[DescTools]{LeveneTest}}.
}
