% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{as_tibble.eda}
\alias{as_tibble.eda}
\title{Convert an EDA object to a tibble.}
\usage{
\method{as_tibble}{eda}(x, newname = "variable", ...)
}
\arguments{
\item{x}{an EDA object, created by \code{\link{eda}}.}

\item{newname}{a name of the data frame variable.}
}
\value{
a data frame.
}
\description{
converts an EDA object to a one-column data frame. The data type and levels are preserved.
}
