% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exploration.R
\name{compare_variables}
\alias{compare_variables}
\title{Compare or correlate features between collectives/cohorts.}
\usage{
compare_variables(
  ...,
  variables,
  split_factor = NULL,
  what = c("test", "distribution", "variance", "correlation", "covariance", "eff_size"),
  types = "t_test",
  exact = TRUE,
  ci = TRUE,
  boot_method = "percentile",
  pub_styled = FALSE,
  signif_digits = 2,
  adj_method = "none",
  simplify_p = TRUE,
  .parallel = FALSE,
  .paropts = furrr::furrr_options(seed = TRUE)
)
}
\arguments{
\item{...}{data frames.}

\item{variables}{a vector with variable names.}

\item{split_factor}{optional, the name of a factor used for splitting the variables of interest into analysis groups.}

\item{what}{the requested analysis: 'test', 'distribution', 'variance', 'correlation' or 'covariance'. Defaults to 'test'.}

\item{types}{a vector with the types of statistical test, see \code{\link{test}}, \code{\link{correlate}} and
\code{\link{covariance}} for details. The vector length must be either one or the length of the 'variables' vector.}

\item{exact}{logical, should exact values for Chi-squared. Mann-Whitney and Wilcoxon test be returned?}

\item{ci}{logical, should confidence intervals for the test effect size be returned?}

\item{boot_method}{indicates how the bootstrap confidence intervals are calculated.
Can be any of 'percentile', 'bca', or 'normality', defaults to 'percentile'.}

\item{pub_styled}{logical, should the output be publication-ready formatted?}

\item{signif_digits}{significant digits used for rounding in the publication-style output.}

\item{adj_method}{the method for adjusting p values for multiple testing, as defined for \code{\link[stats]{p.adjust}},
defaults to 'none'. The adjusted p value is summarized in the 'significance' column, when pub_style output is chosen.}

\item{simplify_p}{logical, should p_values < 0.001 be presented in a p < 0.001 form?}

\item{.parallel}{logical, should the analysis be run in parallel? Experimental, uses the parallel solutions provided by
furrr package.}

\item{.paropts}{an object created by \code{\link[furrr]{furrr_options}}, enabling i.e. provision of globals by the user.}
}
\description{
Compares the given variable between the provided data sets
with a range of statistical sets as described for \code{\link{test}} (statistical testing),
\code{\link{distribution}} (distribution comparison), \code{\link{variance}} (variance comparison),
\code{\link{correlate}} (correlation) or \code{\link{covariance}} (covariance).
}
\details{
in case, the split_factor is provided, only the first of the input data frames will be analyzed.
}
