% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlate_variables}
\alias{correlate_variables}
\title{Pairwise correlation tests.}
\usage{
correlate_variables(
  data,
  variables = NULL,
  type = NULL,
  test_method = c("permutation", "bootstrap"),
  alternative = c("two.sided", "less", "greater"),
  boot_method = c("bca", "percentile"),
  conf_level = 0.95,
  n_iter = 1000,
  adj_method = "none",
  pub_styled = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector with names of variables for which the correlations
will be computed (one against all other). Note, all variables need to be of
the same type (e.g. all numeric). If \code{NULL}, all variables in the table
are correlated.}

\item{type}{\code{NULL} or a string which specifies type of the correlations.
If \code{NULL}, Spearman's correlations are computed for numeric variables, and
Cohen's kappa for factors or integers. Other options are
\code{"pearson"}, \code{"spearman"}, \code{"kendallA"}, \code{"kendallB"}, \code{"xiaA"}, \code{"xiB"}
for numeric variables, and \code{"kappa"} for factors or integers.}

\item{test_method}{method of computing p values: "permutation" (default) or
"bootstrap".}

\item{alternative}{type of the alternative hypothesis concerning correlation
coefficient sign. Ignored if \code{type = "bootstrap"}.}

\item{boot_method}{indicates how the bootstrap confidence intervals are
calculated.
Can be \code{"bca"} for bias-corrected accelerated confidence intervals
(default) or \code{"percentile"} for percentile confidence intervals.}

\item{conf_level}{confidence level used for computation of the
confidence intervals.}

\item{n_iter}{number of permutation or bootstraps.}

\item{adj_method}{the method for adjusting p values for multiple testing,
as defined for \code{\link[stats]{p.adjust}},
defaults to \code{"none"}. The adjusted p value appears in the \code{p_adjusted} column.}

\item{pub_styled}{logical, should the output be returned as a publication-styled
data frame (class \code{\link{etest}})?}

\item{...}{additional arguments passed to \code{\link[fastTest]{f_cor_test}},
\code{\link[fastTest]{f_kappa_test}}, and function \code{\link{etest}} used
to format the results when \code{pub_styled = TRUE}.}
}
\value{
a data frame with the correlation results: either in a native form
of the \code{fastTest} functions or, if \code{pub_styled = TRUE}, in a pre-formatted
form with significance and effect size texts, and ready-to-use plot labels.
}
\description{
Tests for significant correlations of two numeric variables (methods: Pearson,
Spearman, Kendall TauA and TauB, Chatterjee's XiA and XiB) or two integer/factor
variables (Cohen's kappa inter-rater reliability statistic).
P values are computed with the permutation or bootstrap method.
}
\details{
The function uses internally \code{\link[fastTest]{f_cor_test}} and
\code{\link[fastTest]{f_kappa_test}}, for a richer set of correlation options,
please refer to the genuine correlation functions.
If \code{type = NULL}, \code{correlate_variables()} resorts to Spearman's correlations
for numeric variables and Cohen's kappa for factors or integers.
}
