% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{droplevels.eda}
\alias{droplevels.eda}
\title{Dropping empty levels from a factor-type EDA object.}
\usage{
\method{droplevels}{eda}(eda_object)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}
}
\value{
an EDA object.
}
\description{
Removes empty levels from the values of factor-type EDA object.
For numeric-type ones NULL is returned and a warning generated.
}
