% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_factor_methods.R
\name{droplevels.eda}
\alias{droplevels.eda}
\title{Dropping empty levels from a factor-type \code{eda} object.}
\usage{
\method{droplevels}{eda}(x, ...)
}
\arguments{
\item{x}{an \code{\link{eda}} object.}

\item{...}{additional arguments for methods, currently none.}
}
\value{
an \code{\link{eda}} object.
}
\description{
Removes empty levels from the values of factor-type \code{eda} object.
For numeric-type ones \code{NULL} is returned and a warning generated.
}
