% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{eda}
\alias{eda}
\alias{eda.factor}
\alias{eda.numeric}
\alias{eda.character}
\alias{eda.logical}
\alias{eda.data.frame}
\alias{is_eda}
\title{Generate \code{eda} objects.}
\usage{
eda(x, ...)

\method{eda}{factor}(x, .drop = TRUE, ...)

\method{eda}{numeric}(x, ...)

\method{eda}{character}(x, ...)

\method{eda}{logical}(x, ...)

\method{eda}{data.frame}(x, variable, ...)

is_eda(x)
}
\arguments{
\item{x}{and object to be converted to \code{eda} or a data frame with the
variable to be converted to an \code{eda} object.}

\item{...}{additional arguments for methods,}

\item{.drop}{logical, should empty levels of a factor be dropped?
Defaults to \code{TRUE}.}

\item{variable}{a name of data frame's variable.}
}
\value{
an \code{eda} object: an enriched numeric or factor vector, which inherits
many of the numeric or factor canonical methods.
}
\description{
Generates \code{eda} objects from multiple inputs such
as vectors or data frame variables.
\code{eda} objects are internally used vectors with predefined behavior
concerning handling of NAs and descriptive statistics returned by \code{summary()}.
}
