% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{etest}
\alias{etest}
\title{Generate eTest objects.}
\usage{
etest(
  test,
  stat_name = NA,
  stat = NA,
  df1 = NA,
  df2 = NA,
  estimate_name = NA,
  estimate = NA,
  lower_ci = NA,
  upper_ci = NA,
  p_value = NA,
  n = NA
)
}
\arguments{
\item{test}{statistical test name.}

\item{stat_name}{name of the test statistic.}

\item{stat}{value of the test statistic.}

\item{df1}{= degrees of freedom.}

\item{df2}{degrees of freedom.}

\item{estimate_name}{name of the test estimate.}

\item{estimate}{value of the test estimate.}

\item{lower_ci}{value of the lower confidence interval limit.}

\item{upper_ci}{value of the upper confidence interval limit.}

\item{p_value}{p value.}

\item{n}{number of complete observations used for testing.}
}
\value{
an EDA object.
}
\description{
Generates eTest objects storing statistic test results in
a standardized tibble.
}
