% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exploration.R
\name{explore}
\alias{explore}
\title{Explorative data analysis of data frame variables.}
\usage{
explore(
  data,
  variables = names(data),
  what = c("list", "table", "plota", "raw"),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector of variable names.}

\item{what}{the type of output: 'list' returns a list of distribution statistics, 'table' returns
a publication-ready table with the distribution stats, 'plots' returns a list of plots,
'raw' returns a list of EDA objects (\code{\link{eda}}).}

\item{...}{additional arguments passed to \code{\link{summary.eda}} or \code{\link{plot.eda}}.}
}
\value{
as specified by the 'what' argument: a list of stats, a statistic table or a list of plots.
}
\description{
Performs exploratory data analysis for the selected variables of a data frame.
The output may include lists of distribution statistics, a publication-ready table with the distribution statistics
or distribution plots.
}
