% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{explore}
\alias{explore}
\title{Exploratory data analysis of a data frame.}
\usage{
explore(
  data,
  variables = NULL,
  split_factor = NULL,
  what = c("table", "list", "skewness", "kurtosis"),
  .drop = TRUE,
  signif_digits = 2,
  one_table = TRUE,
  total_text = "observations, total, N",
  style = c("full", "mean/SD", "median/IQR"),
  rm_range = FALSE,
  rm_complete = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector of variable names. If \code{NULL} all variables in the
data frame except of the splitting factor will be analyzed.}

\item{split_factor}{optional, the name of a factor used for splitting the
variables of interest into analysis groups.}

\item{what}{the type of output: 'list' returns a list of distribution statistics,
'table' (default) returns a publication-ready data frame with
the distribution stats,
'skewness' and 'kurtosis' return data frames with the requested statistics.}

\item{.drop}{logical, should empty categories of factors in the data frame be removed?}

\item{signif_digits}{significant digits used for rounding in the
publication-style output.}

\item{one_table}{logical, should the output be coerced to a single data frame?
Applies to \code{what} set to 'table', 'skewness' or 'kurtosis', and ignored otherwise.}

\item{total_text}{a character vector (text) in the output's first row with the
total number of observations. Applied only if \code{pub_styled = TRUE}.}

\item{style}{style of the output table, see Details. Defaults to \code{full}.
Used only if \code{what = "table"}.}

\item{rm_range}{logical, should the information on range of numeric variables
be removed from the output data frame. Used only if \code{what = "table"}.}

\item{rm_complete}{logical, should the information on numbers of complete
observations be removed from the output data frame. Used only if \code{what = "table"}.
Useful, when the user analyzes a data frame without missing values.}

\item{...}{additional arguments. Kept for compatibility with earlier versions
of \code{explore()}, which took e.g. \code{pub_styled} argument.}
}
\value{
as specified by the 'what' argument: a list of stats, a data frame
(if \code{one_table = TRUE}) or a list of data frames (if \code{one_table = FALSE})
with descriptive statistics.
}
\description{
Performs exploratory data analysis for the selected variables of a data frame.
The output may include lists of distribution statistics, or
a publication-ready table with the distribution statistics.
}
\details{
The function is way more compact than \code{explore()} in previous version of the
package.
For instance it does not provide a plot output anymore, and skips
the \code{pub_styled} argument which are implicitly assumed to be true, if the user
requests a summary table as an output.
It implements also new features, like total number of observations in the summary
table, and and option to collapse the output into a single data frame.
The \code{style} argument specifies, which descriptive statistics for numeric
features will be displayed in the summary table:
\itemize{
\item \code{"full"}: means, standard deviations, medians, and interquartile ranges
}

*\code{"mean/SD"}: means with standard deviations
\itemize{
\item \code{"median/IQR"}: medians with interquartile ranges
}
}
