% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{length.eda}
\alias{length.eda}
\title{Length of EDA objects.}
\usage{
\method{length}{eda}(x)
}
\arguments{
\item{x}{an EDA object, created by \code{\link{eda}}.}
}
\value{
a numeric value.
}
\description{
returns length of the EDA object's value vector.
}
