% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{na.omit.eda}
\alias{na.omit.eda}
\title{NA removal from EDA objects.}
\usage{
\method{na.omit}{eda}(object, ...)
}
\arguments{
\item{object}{an EDA object, created by \code{\link{eda}}.}
}
\value{
an EDA objects without NAs.
}
\description{
removes NAs from the EDA object.
}
