% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{nobs.eda}
\alias{nobs.eda}
\title{Number of (complete) observations in an EDA object.}
\usage{
\method{nobs}{eda}(eda_object)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}
}
\value{
a tibble with the number of all and complete observations.
}
\description{
counts the number of all (complete if na.rm is set to TRUE)
observations in the EDA object.
}
