% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_general_methods.R
\name{plot.eda}
\alias{plot.eda}
\title{Plot distribution.}
\usage{
\method{plot}{eda}(x, type = NULL, ...)
}
\arguments{
\item{x}{\code{\link{eda}} object.}

\item{type}{plot type: "stack", "bar", "violin", "box", "histogram", "density",
or "qq".
If not provided, a stack plot is returned for factors and a violin plot is
returned for numeric objects.}

\item{...}{additional arguments passed to internal plotting functions
(bar and stack plots: \code{\link{plot_factor}}, violin and box plots:
\code{\link{plot_numeric}}, histograms: \code{\link{plot_histogram}}, QQ plots:
\code{\link{plot_qq}}). They specify, among others, colors and opacity of
points and shapes, point sizes, text labels, and \code{ggplot} themes.}
}
\value{
a \code{ggplot} object.
}
\description{
Distribution plots for \code{eda} objects storing factors (bar and stack plots
with frequencies of observations in the categories) and numeric values
(violin and box plots, histograms, density plots,
and quantile - quantile/QQ plots).
}
