% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot.etest}
\alias{plot.etest}
\title{Plot effect size for the eTest object.}
\usage{
\method{plot}{etest}(
  x,
  point_alpha = 0.5,
  point_hjitter = 0,
  point_wjitter = 0,
  point_size = 2,
  point_color = c("gray60", "coral2"),
  plot_title = NULL,
  plot_subtitle = NULL,
  cust_theme = ggplot2::theme_classic(),
  show_labels = c("none", "all", "signif"),
  txt_size = 2.75,
  ...
)
}
\arguments{
\item{x}{eTest object, created e.g. by \code{\link{compare_variables}}.}

\item{point_alpha}{alpha of the plot points.}

\item{point_hjitter}{point jitter height.}

\item{point_wjitter}{point jitter width.}

\item{point_size}{size of the points in the plots.}

\item{point_color}{a pair of valid R colors: 1) for non-significant and 2) for significant effects.}

\item{plot_title}{text to be displayed in the plot title.}

\item{plot_subtitle}{text to be displayed in the plot subtitle.}

\item{cust_theme}{custom ggplot2 theme.}

\item{show_labels}{which points should be labeled? 'none' hides labels (default), 'all': labels all points,
'signif': labels only the significant effects with the variable name.}

\item{txt_size}{size of the label text.}

\item{...}{additional arguments passed to \code{\link[ggrepel]{geom_text_repel}}.}
}
\description{
Plots effect size statistic and the adjusted significance for
the given eTest object storing results of a statistic test. Plot tag contains
(range) n numbers of complete observations.
}
