% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot variable distribution.}
\usage{
plot_variable(data, variable, split_factor, type = NULL, ...)
}
\arguments{
\item{data}{a data frame.}

\item{variable}{name of the data frame's variable whose distribution will be
shown in the plot. Provided as a string.}

\item{split_factor}{name of the data frame's variable, whose categories will
be used to split values of \code{variable} in the plot. Provided as a string.}

\item{type}{plot type: "stack", "bar", "violin", "box", "paired",
"histogram", "density", or "qq".
If not provided, a stack plot is returned for factors and a violin plot is
returned for numeric objects.}

\item{...}{additional arguments passed to internal plotting functions
(bar and stack plots: \code{\link{plot_df_factor}}; violin, box and paired
plots: \code{\link{plot_df_numeric}}, histograms: \code{\link{plot_df_histogram}},
QQ plots: \code{\link{plot_df_qq}}).
They specify, among others, colors and opacity of
points and shapes, point sizes, text labels, and \code{ggplot} themes.}
}
\value{
a \code{ggplot} object.
}
\description{
\code{ggplot} plots with distribution of a variable in analysis groups.
For factor variables, stack, bar, and dot plots with counts or percentages of
observations in the analysis groups.
For numeric variables, violin and box plots, paired/matched plots ("before/after"),
histograms, density plots, and quantile - quantile (QQ) plots are available.
}
\details{
If \code{type = "paired"}, a before-after plot is generated and its assumed that
the plotting data are already arranged by the blocking or identified variable.
}
