% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exploration.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot a variable in two or more data sets objects.}
\usage{
plot_variable(
  ...,
  variable,
  split_factor = NULL,
  data_names = NULL,
  type = c("default", "bar", "violin", "box", "hist", "correlation", "paired", "stack"),
  scale = c("none", "fraction", "percent"),
  point_alpha = 0.5,
  point_hjitter = 0.05,
  point_wjitter = 0.1,
  point_color = "steelblue",
  point_size = 2,
  line_color = "black",
  line_alpha = 0.25,
  cust_theme = ggplot2::theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  show_trend = TRUE,
  show_labels = TRUE,
  geom_label = TRUE,
  signif_digits = 2,
  txt_size = 2.75,
  bins = NULL,
  facet_hist = c("none", "horizontal", "vertical")
)
}
\arguments{
\item{...}{data sets.}

\item{variable}{variable name.}

\item{split_factor}{optional, the name of a factor used for splitting the variables of interest into analysis groups.}

\item{data_names}{a vector with names of the data sets.}

\item{type}{type of the plot. 'default' plots violin for numeric variables and bars for factors.
'bar', 'bubble' and 'stack' (stacked bar plot) are available for factors.
'violin', 'box', 'hist', 'correlation' and 'paired'
are available for numeric variables.}

\item{scale}{the feature to be presented in factor bar plots. 'none' plots counts, 'percent' plots percentages,
'fraction' presents fraction fo complete observations.}

\item{point_alpha}{alpha of the plot points.}

\item{point_hjitter}{point jitter height.}

\item{point_wjitter}{point jitter width.}

\item{point_color}{color of the points in the correlation plot.}

\item{point_size}{size of the points in the plots.}

\item{line_color}{color of the trend line in the correlation plots or the connecting lines in the paired plots.}

\item{line_alpha}{opacity of the connecting lines in the paired plot.}

\item{cust_theme}{custom ggplot2 theme.}

\item{plot_title}{text to be presented in the plot title.}

\item{plot_subtitle}{text to be presented in the plot subtitle.}

\item{x_lab}{text to be presented in the X axis title.}

\item{y_lab}{text to be presented in the Y axis title.}

\item{show_trend}{logical, should a trend line with 95\% confidence intervals be presented in the correlation plots?}

\item{show_labels}{logical, should labels with count numbers, percentages or fractions be presented in bar plots?}

\item{geom_label}{logical, should the text in the stacked bar plot be
presented as a ggplot's geom_label?}

\item{signif_digits}{significant digits used for the label value rounding.}

\item{txt_size}{size of the text label.}

\item{bins}{bin number, passed to \code{\link[ggplot2]{histogram}}.}

\item{facet_hist}{'none': histograms are overlaid, 'horizontal': horizontal or 'vertical': vertical faceting.}
}
\description{
Plots values of two or more EDA objects as a classical bar, violin,
boxplot or correlation point plot representation.
}
\details{
the particular data sets representations are color coded. In case, the split_factor is provided, only the first
of the input data frames will be plotted.
}
