% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{print.eda}
\alias{print.eda}
\title{Printing EDA objects.}
\usage{
\method{print}{eda}(eda_object)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}
}
\value{
nothing, called for its side effects.
}
\description{
print an EDA object.
}
