% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{print.eda}
\alias{print.eda}
\alias{print.etest}
\title{Appearance/printing of \code{eda} and \code{etest} objects}
\usage{
\method{print}{eda}(x, ...)

\method{print}{etest}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{arguments for methods.}
}
\value{
invisibly returns the object.
}
\description{
Printing methods.
}
