% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_methods.R
\name{quantile.eda}
\alias{quantile.eda}
\title{Quantiles of an EDA object.}
\usage{
\method{quantile}{eda}(eda_object, probs = c(0, 0.25, 0.5, 0.75, 1), plain = FALSE)
}
\arguments{
\item{eda_object}{an EDA object, created by \code{\link{eda}}.}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{plain}{logical, should the output be coerced to a single vector?}
}
\value{
a tibble or vector with the requested quantile values.
}
\description{
Gets quantiles of a numeric-type EDA object. For factor-type ones
NULL is returned and a warning generated. NAs are removed before computation.
}
