% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_methods.R
\name{scale.eda}
\alias{scale.eda}
\title{Normalize an EDA object.}
\usage{
\method{scale}{eda}(x, center = TRUE, scale = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{an EDA object, created by \code{\link{eda}}.}

\item{center}{see: \code{\link[base]{scale}}.}

\item{scale}{see: \code{\link[base]{scale}}.}

\item{na.rm}{logical, should NAs be removed prior to scaling?}
}
\description{
Normalizes values of an EDA object.
}
\details{
A wrapper around \code{\link[base]{scale}}.
}
