% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_general_methods.R
\name{summary.eda}
\alias{summary.eda}
\title{Summary of \code{eda} objects.}
\usage{
\method{summary}{eda}(object, pub_styled = FALSE, signif_digits = 2, ...)
}
\arguments{
\item{object}{\code{\link{eda}} object.}

\item{pub_styled}{logical, should the output be publication-ready formatted?
See the details.}

\item{signif_digits}{number of significant digits used for rounding in
the publication-style output.}

\item{...}{additional arguments for methods.}
}
\description{
Summary of numeric and factor \code{eda} objects.
}
\details{
The \code{pub_styled} argument allows for generating a 'nicer' output
accepted by multiple journals.
A tibble with the column 'statistic' is returned.
}
